Module wasmedge_sys::ffi

Foreign function interfaces generated from WasmEdge C-API.

Structs

WasmEdge_ASTModuleContext

WasmEdge_Async

WasmEdge_CallingFrameContext

WasmEdge_CompilerContext

WasmEdge_ConfigureContext

WasmEdge_ExecutorContext

WasmEdge_ExportTypeContext

WasmEdge_FunctionInstanceContext

WasmEdge_FunctionTypeContext

WasmEdge_GlobalInstanceContext

WasmEdge_GlobalTypeContext

WasmEdge_ImportTypeContext

WasmEdge_Limit

Struct of WASM limit.

WasmEdge_LoaderContext

WasmEdge_MemoryInstanceContext

WasmEdge_MemoryTypeContext

WasmEdge_ModuleDescriptor

Module descriptor for plugins.

WasmEdge_ModuleInstanceContext

WasmEdge_PluginContext

WasmEdge_PluginDescriptor

Plugin descriptor for plugins.

WasmEdge_PluginVersionData

Version data for plugins.

WasmEdge_ProgramOption

Program option for plugins.

WasmEdge_Result

WasmEdge result struct.

WasmEdge_StatisticsContext

WasmEdge_StoreContext

WasmEdge_String

WasmEdge string struct.

WasmEdge_TableInstanceContext

WasmEdge_TableTypeContext

WasmEdge_VMContext

WasmEdge_ValidatorContext

WasmEdge_Value

WasmEdge WASM value struct.

__fsid_t

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WASMEDGE_VERSION

WASMEDGE_VERSION_MAJOR

WASMEDGE_VERSION_MINOR

WASMEDGE_VERSION_PATCH

WINT_MAX

WINT_MIN

WasmEdge_CompilerOptimizationLevel_O0

WasmEdge_CompilerOptimizationLevel_O1

WasmEdge_CompilerOptimizationLevel_O2

WasmEdge_CompilerOptimizationLevel_O3

WasmEdge_CompilerOptimizationLevel_Os

WasmEdge_CompilerOptimizationLevel_Oz

WasmEdge_CompilerOutputFormat_Native

WasmEdge_CompilerOutputFormat_Wasm

WasmEdge_ErrCategory_UserLevelError

WasmEdge_ErrCategory_WASM

WasmEdge_ErrCode_AOTDisabled

WasmEdge_ErrCode_ConstExprRequired

WasmEdge_ErrCode_CostLimitExceeded

WasmEdge_ErrCode_DataCountRequired

WasmEdge_ErrCode_DataSegDoesNotFit

WasmEdge_ErrCode_DivideByZero

WasmEdge_ErrCode_DupExportName

WasmEdge_ErrCode_ENDCodeExpected

WasmEdge_ErrCode_ElemSegDoesNotFit

WasmEdge_ErrCode_ExpectSharedMemory

WasmEdge_ErrCode_ExpectedZeroByte

WasmEdge_ErrCode_FuncNotFound

WasmEdge_ErrCode_FuncSigMismatch

WasmEdge_ErrCode_HostFuncError

WasmEdge_ErrCode_IllegalGrammar

WasmEdge_ErrCode_IllegalOpCode

WasmEdge_ErrCode_IllegalPath

WasmEdge_ErrCode_ImmutableGlobal

WasmEdge_ErrCode_IncompatibleDataCount

WasmEdge_ErrCode_IncompatibleFuncCode

WasmEdge_ErrCode_IncompatibleImportType

WasmEdge_ErrCode_IndirectCallTypeMismatch

WasmEdge_ErrCode_InstrTypeMismatch

WasmEdge_ErrCode_IntegerOverflow

WasmEdge_ErrCode_IntegerTooLarge

WasmEdge_ErrCode_IntegerTooLong

WasmEdge_ErrCode_Interrupted

WasmEdge_ErrCode_IntrinsicsTableNotFound

WasmEdge_ErrCode_InvalidAlignment

WasmEdge_ErrCode_InvalidConvToInt

WasmEdge_ErrCode_InvalidDataIdx

WasmEdge_ErrCode_InvalidElemIdx

WasmEdge_ErrCode_InvalidFuncIdx

WasmEdge_ErrCode_InvalidFuncTypeIdx

WasmEdge_ErrCode_InvalidGlobalIdx

WasmEdge_ErrCode_InvalidLabelIdx

WasmEdge_ErrCode_InvalidLaneIdx

WasmEdge_ErrCode_InvalidLimit

WasmEdge_ErrCode_InvalidLocalIdx

WasmEdge_ErrCode_InvalidMemPages

WasmEdge_ErrCode_InvalidMemoryIdx

WasmEdge_ErrCode_InvalidMut

WasmEdge_ErrCode_InvalidRefIdx

WasmEdge_ErrCode_InvalidResultArity

WasmEdge_ErrCode_InvalidStartFunc

WasmEdge_ErrCode_InvalidTableIdx

WasmEdge_ErrCode_JunkSection

WasmEdge_ErrCode_LengthOutOfBounds

WasmEdge_ErrCode_MalformedElemType

WasmEdge_ErrCode_MalformedExportKind

WasmEdge_ErrCode_MalformedImportKind

WasmEdge_ErrCode_MalformedMagic

WasmEdge_ErrCode_MalformedRefType

WasmEdge_ErrCode_MalformedSection

WasmEdge_ErrCode_MalformedUTF8

WasmEdge_ErrCode_MalformedValType

WasmEdge_ErrCode_MalformedVersion

WasmEdge_ErrCode_MemoryOutOfBounds

WasmEdge_ErrCode_ModuleNameConflict

WasmEdge_ErrCode_MultiMemories

WasmEdge_ErrCode_MultiTables

WasmEdge_ErrCode_NotValidated

WasmEdge_ErrCode_ReadError

WasmEdge_ErrCode_RefTypeMismatch

WasmEdge_ErrCode_RuntimeError

WasmEdge_ErrCode_SectionSizeMismatch

WasmEdge_ErrCode_SharedMemoryNoMax

WasmEdge_ErrCode_Success

WasmEdge_ErrCode_TableOutOfBounds

WasmEdge_ErrCode_Terminated

WasmEdge_ErrCode_TooManyLocals

WasmEdge_ErrCode_TypeCheckFailed

WasmEdge_ErrCode_UnalignedAtomicAccess

WasmEdge_ErrCode_UndefinedElement

WasmEdge_ErrCode_UnexpectedEnd

WasmEdge_ErrCode_UninitializedElement

WasmEdge_ErrCode_UnknownImport

WasmEdge_ErrCode_Unreachable

WasmEdge_ErrCode_UserDefError

WasmEdge_ErrCode_WrongInstanceAddress

WasmEdge_ErrCode_WrongInstanceIndex

WasmEdge_ErrCode_WrongVMWorkflow

WasmEdge_ExternalType_Function

WasmEdge_ExternalType_Global

WasmEdge_ExternalType_Memory

WasmEdge_ExternalType_Table

WasmEdge_HostRegistration_Wasi

WasmEdge_Mutability_Const

WasmEdge_Mutability_Var

WasmEdge_NumType_F32

WasmEdge_NumType_F64

WasmEdge_NumType_I32

WasmEdge_NumType_I64

WasmEdge_NumType_V128

WasmEdge_Plugin_CurrentAPIVersion

WasmEdge_ProgramOptionType_Double

Boolean value.

WasmEdge_ProgramOptionType_Float

Boolean value.

WasmEdge_ProgramOptionType_Int8

Boolean value.

WasmEdge_ProgramOptionType_Int16

Boolean value.

WasmEdge_ProgramOptionType_Int32

Boolean value.

WasmEdge_ProgramOptionType_Int64

Boolean value.

WasmEdge_ProgramOptionType_None

No option value.

WasmEdge_ProgramOptionType_String

WasmEdge_String.

WasmEdge_ProgramOptionType_Toggle

Boolean value.

WasmEdge_ProgramOptionType_UInt8

Boolean value.

WasmEdge_ProgramOptionType_UInt16

Boolean value.

WasmEdge_ProgramOptionType_UInt32

Boolean value.

WasmEdge_ProgramOptionType_UInt64

Boolean value.

WasmEdge_Proposal_Annotations

WasmEdge_Proposal_BulkMemoryOperations

WasmEdge_Proposal_ExceptionHandling

WasmEdge_Proposal_ExtendedConst

WasmEdge_Proposal_FunctionReferences

WasmEdge_Proposal_ImportExportMutGlobals

WasmEdge_Proposal_Memory64

WasmEdge_Proposal_MultiMemories

WasmEdge_Proposal_MultiValue

WasmEdge_Proposal_NonTrapFloatToIntConversions

WasmEdge_Proposal_ReferenceTypes

WasmEdge_Proposal_SIMD

WasmEdge_Proposal_SignExtensionOperators

WasmEdge_Proposal_TailCall

WasmEdge_Proposal_Threads

WasmEdge_RefType_ExternRef

WasmEdge_RefType_FuncRef

WasmEdge_ValType_ExternRef

WasmEdge_ValType_F32

WasmEdge_ValType_F64

WasmEdge_ValType_FuncRef

WasmEdge_ValType_I32

WasmEdge_ValType_I64

WasmEdge_ValType_V128

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

true_

Functions

WasmEdge_ASTModuleDelete⚠

Deletion of the WasmEdge_ASTModuleContext.

WasmEdge_ASTModuleListExports⚠

List the exports of the AST module.

WasmEdge_ASTModuleListExportsLength⚠

Get the length of exports list of the AST module.

WasmEdge_ASTModuleListImports⚠

List the imports of the AST module.

WasmEdge_ASTModuleListImportsLength⚠

Get the length of imports list of the AST module.

WasmEdge_AsyncCancel⚠

Cancel a WasmEdge_Async execution.

WasmEdge_AsyncDelete⚠

Deletion of the WasmEdge_Async.

WasmEdge_AsyncGet⚠

Wait and get the result of WasmEdge_Async execution.

WasmEdge_AsyncGetReturnsLength⚠

Wait and get the return list length of the WasmEdge_Async execution.

WasmEdge_AsyncWait⚠

Wait a WasmEdge_Async execution.

WasmEdge_AsyncWaitFor⚠

Wait a WasmEdge_Async execution with timeout.

WasmEdge_CallingFrameGetExecutor⚠

Get the executor context from the current calling frame.

WasmEdge_CallingFrameGetMemoryInstance⚠

Get the memory instance by index from the module instance of the current calling frame.

WasmEdge_CallingFrameGetModuleInstance⚠

Get the module instance of the current calling frame.

WasmEdge_CompilerCompile⚠

Compile the input WASM from the file path.

WasmEdge_CompilerCompileFromBuffer⚠

Compile the input WASM from the given buffer.

WasmEdge_CompilerCreate⚠

Creation of the WasmEdge_CompilerContext.

WasmEdge_CompilerDelete⚠

Deletion of the WasmEdge_CompilerContext.

WasmEdge_ConfigureAddHostRegistration⚠

Add a built-in host registration setting into WasmEdge_ConfigureContext.

WasmEdge_ConfigureAddProposal⚠

Add a proposal setting into the WasmEdge_ConfigureContext.

WasmEdge_ConfigureCompilerGetOptimizationLevel⚠

Get the optimization level of the AOT compiler.

WasmEdge_ConfigureCompilerGetOutputFormat⚠

Get the output binary format of the AOT compiler.

WasmEdge_ConfigureCompilerIsDumpIR⚠

Get the dump IR option of the AOT compiler.

WasmEdge_ConfigureCompilerIsGenericBinary⚠

Get the generic binary option of the AOT compiler.

WasmEdge_ConfigureCompilerIsInterruptible⚠

Get the interruptible option of the AOT compiler.

WasmEdge_ConfigureCompilerSetDumpIR⚠

Set the dump IR option of the AOT compiler.

WasmEdge_ConfigureCompilerSetGenericBinary⚠

Set the generic binary option of the AOT compiler.

WasmEdge_ConfigureCompilerSetInterruptible⚠

Set the interruptible option of the AOT compiler.

WasmEdge_ConfigureCompilerSetOptimizationLevel⚠

Set the optimization level of the AOT compiler.

WasmEdge_ConfigureCompilerSetOutputFormat⚠

Set the output binary format of the AOT compiler.

WasmEdge_ConfigureCreate⚠

Creation of the WasmEdge_ConfigureContext.

WasmEdge_ConfigureDelete⚠

Deletion of the WasmEdge_ConfigureContext.

WasmEdge_ConfigureGetMaxMemoryPage⚠

Get the setting of the page limit of memory instances.

WasmEdge_ConfigureHasHostRegistration⚠

Check if a built-in host registration setting exists in the WasmEdge_ConfigureContext or not.

WasmEdge_ConfigureHasProposal⚠

Check if a proposal setting exists in the WasmEdge_ConfigureContext or not.

WasmEdge_ConfigureIsAllowAFUNIX⚠

Get the AllowAFUNIX option.

WasmEdge_ConfigureIsForceInterpreter⚠

Get the force interpreter mode execution option.

WasmEdge_ConfigureRemoveHostRegistration⚠

Remove a built-in host registration setting in the WasmEdge_ConfigureContext.

WasmEdge_ConfigureRemoveProposal⚠

Remove a proposal setting in the WasmEdge_ConfigureContext.

WasmEdge_ConfigureSetAllowAFUNIX⚠

Set the option of enabling/disabling AF_UNIX support in the WASI socket.

WasmEdge_ConfigureSetForceInterpreter⚠

Set the force interpreter mode execution option.

WasmEdge_ConfigureSetMaxMemoryPage⚠

Set the page limit of memory instances.

WasmEdge_ConfigureStatisticsIsCostMeasuring⚠

Get the cost measuring option for the statistics.

WasmEdge_ConfigureStatisticsIsInstructionCounting⚠

Get the instruction counting option for the statistics.

WasmEdge_ConfigureStatisticsIsTimeMeasuring⚠

Get the time measuring option for the statistics.

WasmEdge_ConfigureStatisticsSetCostMeasuring⚠

Set the cost measuring option for the statistics.

WasmEdge_ConfigureStatisticsSetInstructionCounting⚠

Set the instruction counting option for the statistics.

WasmEdge_ConfigureStatisticsSetTimeMeasuring⚠

Set the time measuring option for the statistics.

WasmEdge_Driver_Compiler⚠

Entrypoint for the compiler tool.

WasmEdge_Driver_Tool⚠

Entrypoint for the runtime tool.

WasmEdge_Driver_UniTool⚠

Entrypoint for the unified tool.

WasmEdge_ExecutorAsyncInvoke⚠

Asynchronous invoke a WASM function by the function instance.

WasmEdge_ExecutorCreate⚠

Creation of the WasmEdge_ExecutorContext.

WasmEdge_ExecutorDelete⚠

Deletion of the WasmEdge_ExecutorContext.

WasmEdge_ExecutorExperimentalRegisterPostHostFunction⚠

Register a host function that will be invoked after executing any host functions.

WasmEdge_ExecutorExperimentalRegisterPreHostFunction⚠

Register a host function that will be invoked before executing any host functions.

WasmEdge_ExecutorInstantiate⚠

Instantiate an AST Module into a module instance.

WasmEdge_ExecutorInvoke⚠

Invoke a WASM function by the function instance.

WasmEdge_ExecutorRegister⚠

Instantiate an AST Module into a named module instance and link into store.

WasmEdge_ExecutorRegisterImport⚠

Register a module instance into a store with exporting its module name.

WasmEdge_ExportTypeGetExternalName⚠

Get the external name from an export type.

WasmEdge_ExportTypeGetExternalType⚠

Get the external type from an export type.

WasmEdge_ExportTypeGetFunctionType⚠

Get the external value (which is function type) from an export type.

WasmEdge_ExportTypeGetGlobalType⚠

Get the external value (which is global type) from an export type.

WasmEdge_ExportTypeGetMemoryType⚠

Get the external value (which is memory type) from an export type.

WasmEdge_ExportTypeGetTableType⚠

Get the external value (which is table type) from an export type.

WasmEdge_FunctionInstanceCreate⚠

Creation of the WasmEdge_FunctionInstanceContext for host functions.

WasmEdge_FunctionInstanceCreateBinding⚠

Creation of the WasmEdge_FunctionInstanceContext for host functions.

WasmEdge_FunctionInstanceDelete⚠

Deletion of the WasmEdge_FunctionInstanceContext.

WasmEdge_FunctionInstanceGetData⚠

Get the function data field of the function instance.

WasmEdge_FunctionInstanceGetFunctionType⚠

Get the function type context of the function instance.

WasmEdge_FunctionTypeCreate⚠

Creation of the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeDelete⚠

Deletion of the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetParameters⚠

Get the parameter types list from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetParametersLength⚠

Get the parameter types list length from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetReturns⚠

Get the return types list from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetReturnsLength⚠

Get the return types list length from the WasmEdge_FunctionTypeContext.

WasmEdge_GlobalInstanceCreate⚠

Creation of the WasmEdge_GlobalInstanceContext.

WasmEdge_GlobalInstanceDelete⚠

Deletion of the WasmEdge_GlobalInstanceContext.

WasmEdge_GlobalInstanceGetGlobalType⚠

Get the global type context from a global instance.

WasmEdge_GlobalInstanceGetValue⚠

Get the value from a global instance.

WasmEdge_GlobalInstanceSetValue⚠

Set the value from a global instance.

WasmEdge_GlobalTypeCreate⚠

Creation of the WasmEdge_GlobalTypeContext.

WasmEdge_GlobalTypeDelete⚠

Deletion of the WasmEdge_GlobalTypeContext.

WasmEdge_GlobalTypeGetMutability⚠

Get the mutability from a global type.

WasmEdge_GlobalTypeGetValType⚠

Get the value type from a global type.

WasmEdge_ImportTypeGetExternalName⚠

Get the external name from an import type.

WasmEdge_ImportTypeGetExternalType⚠

Get the external type from an import type.

WasmEdge_ImportTypeGetFunctionType⚠

Get the external value (which is function type) from an import type.

WasmEdge_ImportTypeGetGlobalType⚠

Get the external value (which is global type) from an import type.

WasmEdge_ImportTypeGetMemoryType⚠

Get the external value (which is memory type) from an import type.

WasmEdge_ImportTypeGetModuleName⚠

Get the module name from an import type.

WasmEdge_ImportTypeGetTableType⚠

Get the external value (which is table type) from an import type.

WasmEdge_LimitIsEqual⚠

Compare the two WasmEdge_Limit objects.

WasmEdge_LoaderCreate⚠

Creation of the WasmEdge_LoaderContext.

WasmEdge_LoaderDelete⚠

Deletion of the WasmEdge_LoaderContext.

WasmEdge_LoaderParseFromBuffer⚠

Load and parse the WASM module from a buffer into WasmEdge_ASTModuleContext.

WasmEdge_LoaderParseFromFile⚠

Load and parse the WASM module from a WASM file into a WasmEdge_ASTModuleContext.

WasmEdge_LogOff⚠

Set the logging system off.

WasmEdge_LogSetDebugLevel⚠

Set the logging system to filter to debug level.

WasmEdge_LogSetErrorLevel⚠

Set the logging system to filter to error level.

WasmEdge_MemoryInstanceCreate⚠

Creation of the WasmEdge_MemoryInstanceContext.

WasmEdge_MemoryInstanceDelete⚠

Deletion of the WasmEdge_MemoryInstanceContext.

WasmEdge_MemoryInstanceGetData⚠

Copy the data to the output buffer from a memory instance.

WasmEdge_MemoryInstanceGetMemoryType⚠

Get the memory type context from a memory instance.

WasmEdge_MemoryInstanceGetPageSize⚠

Get the current page size (64 KiB of each page) of a memory instance.

WasmEdge_MemoryInstanceGetPointer⚠

Get the data pointer in a memory instance.

WasmEdge_MemoryInstanceGetPointerConst⚠

Get the const data pointer in a const memory instance.

WasmEdge_MemoryInstanceGrowPage⚠

Grow a memory instance with a page size.

WasmEdge_MemoryInstanceSetData⚠

Copy the data into a memory instance from the input buffer.

WasmEdge_MemoryTypeCreate⚠

Creation of the WasmEdge_MemoryTypeContext.

WasmEdge_MemoryTypeDelete⚠

Deletion of the WasmEdge_MemoryTypeContext.

WasmEdge_MemoryTypeGetLimit⚠

Get the limit from a memory type.

WasmEdge_ModuleInstanceAddFunction⚠

Add a function instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddGlobal⚠

Add a global instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddMemory⚠

Add a memory instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddTable⚠

Add a table instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceCreate⚠

Creation of the WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceCreateWASI⚠

Creation of the WasmEdge_ModuleInstanceContext for the WASI specification.

WasmEdge_ModuleInstanceCreateWithData⚠

Creation of the WasmEdge_ModuleInstanceContext with host data.

WasmEdge_ModuleInstanceDelete⚠

Deletion of the WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceFindFunction⚠

Get the exported function instance context of a module instance.

WasmEdge_ModuleInstanceFindGlobal⚠

Get the exported global instance context of a module instance.

WasmEdge_ModuleInstanceFindMemory⚠

Get the exported memory instance context of a module instance.

WasmEdge_ModuleInstanceFindTable⚠

Get the exported table instance context of a module instance.

WasmEdge_ModuleInstanceGetHostData⚠

Get the host data set into the module instance when creating.

WasmEdge_ModuleInstanceGetModuleName⚠

Get the export module name of a module instance.

WasmEdge_ModuleInstanceInitWASI⚠

Initialize the WasmEdge_ModuleInstanceContext for the WASI specification.

WasmEdge_ModuleInstanceInitWasmEdgeProcess⚠

Initialize the WasmEdge_ModuleInstanceContext for the wasmedge_process specification.

WasmEdge_ModuleInstanceListFunction⚠

List the exported function names of a module instance.

WasmEdge_ModuleInstanceListFunctionLength⚠

Get the length of exported function list of a module instance.

WasmEdge_ModuleInstanceListGlobal⚠

List the exported global names of a module instance.

WasmEdge_ModuleInstanceListGlobalLength⚠

Get the length of exported global list of a module instance.

WasmEdge_ModuleInstanceListMemory⚠

List the exported memory names of a module instance.

WasmEdge_ModuleInstanceListMemoryLength⚠

Get the length of exported memory list of a module instance.

WasmEdge_ModuleInstanceListTable⚠

List the exported table names of a module instance.

WasmEdge_ModuleInstanceListTableLength⚠

Get the length of exported table list of a module instance.

WasmEdge_ModuleInstanceWASIGetExitCode⚠

Get the WASI exit code.

WasmEdge_ModuleInstanceWASIGetNativeHandler⚠

Get the native handler from the WASI mapped FD/Handler.

WasmEdge_PluginCreateModule⚠

Create the module instance in the plug-in by the module name.

WasmEdge_PluginFind⚠

Find the loaded plug-in context by name.

WasmEdge_PluginGetPluginName⚠

Get the plug-in name of the plug-in context.

WasmEdge_PluginInitWASINN⚠

Initialize the wasi_nn plug-in.

WasmEdge_PluginListModule⚠

List the modules in the plug-in context with their names.

WasmEdge_PluginListModuleLength⚠

Get the length of module list in the plug-in context.

WasmEdge_PluginListPlugins⚠

List the loaded plug-ins with their names.

WasmEdge_PluginListPluginsLength⚠

Get the length of loaded plug-in list.

WasmEdge_PluginLoadFromPath⚠

Load the plugin with the given file or directory.

WasmEdge_PluginLoadWithDefaultPaths⚠

Load plugins with the default search paths.

WasmEdge_Plugin_GetDescriptor⚠

Implement by plugins for returning the plugin descriptor.

WasmEdge_ResultGen⚠

Generate the result with code.

WasmEdge_ResultGetCategory⚠

Get the error category.

WasmEdge_ResultGetCode⚠

Get the result code.

WasmEdge_ResultGetMessage⚠

Get the result message.

WasmEdge_ResultOK⚠

Check the result is a success or not.

WasmEdge_StatisticsClear⚠

Clear all data in the WasmEdge_StatisticsContext.

WasmEdge_StatisticsCreate⚠

Creation of the WasmEdge_StatisticsContext.

WasmEdge_StatisticsDelete⚠

Deletion of the WasmEdge_StatisticsContext.

WasmEdge_StatisticsGetInstrCount⚠

Get the instruction count in execution.

WasmEdge_StatisticsGetInstrPerSecond⚠

Get the instruction count per second in execution.

WasmEdge_StatisticsGetTotalCost⚠

Get the total cost in execution.

WasmEdge_StatisticsSetCostLimit⚠

Set the cost limit in execution.

WasmEdge_StatisticsSetCostTable⚠

Set the costs of instructions.

WasmEdge_StoreCreate⚠

Creation of the WasmEdge_StoreContext.

WasmEdge_StoreDelete⚠

Deletion of the WasmEdge_StoreContext.

WasmEdge_StoreFindModule⚠

Get the module instance context by the module name.

WasmEdge_StoreListModule⚠

List the registered module names.

WasmEdge_StoreListModuleLength⚠

Get the length of registered module list in store.

WasmEdge_StringCopy⚠

Copy the content of WasmEdge_String object to the buffer.

WasmEdge_StringCreateByBuffer⚠

Creation of the WasmEdge_String with the buffer and its length.

WasmEdge_StringCreateByCString⚠

Creation of the WasmEdge_String with the C string.

WasmEdge_StringDelete⚠

Deletion of the WasmEdge_String.

WasmEdge_StringIsEqual⚠

Compare the two WasmEdge_String objects.

WasmEdge_StringWrap⚠

Create the WasmEdge_String wraps to the buffer.

WasmEdge_TableInstanceCreate⚠

Creation of the WasmEdge_TableInstanceContext.

WasmEdge_TableInstanceDelete⚠

Deletion of the WasmEdge_TableInstanceContext.

WasmEdge_TableInstanceGetData⚠

Get the reference value in a table instance.

WasmEdge_TableInstanceGetSize⚠

Get the size of a table instance.

WasmEdge_TableInstanceGetTableType⚠

Get the table type context from a table instance.

WasmEdge_TableInstanceGrow⚠

Grow a table instance with a size.

WasmEdge_TableInstanceSetData⚠

Set the reference value into a table instance.

WasmEdge_TableTypeCreate⚠

Creation of the WasmEdge_TableTypeContext.

WasmEdge_TableTypeDelete⚠

Deletion of the WasmEdge_TableTypeContext.

WasmEdge_TableTypeGetLimit⚠

Get the limit from a table type.

WasmEdge_TableTypeGetRefType⚠

Get the reference type from a table type.

WasmEdge_VMAsyncExecute⚠

Asynchronous invoke a WASM function by name.

WasmEdge_VMAsyncExecuteRegistered⚠

Asynchronous invoke a WASM function by its module name and function name.

WasmEdge_VMAsyncRunWasmFromASTModule⚠

Instantiate the WASM module from a WasmEdge AST Module and asynchronous invoke a function by name.

WasmEdge_VMAsyncRunWasmFromBuffer⚠

Instantiate the WASM module from a buffer and asynchronous invoke a function by name.

WasmEdge_VMAsyncRunWasmFromFile⚠

Instantiate the WASM module from a WASM file and asynchronous invoke a function by name.

WasmEdge_VMCleanup⚠

Reset of WasmEdge_VMContext.

WasmEdge_VMCreate⚠

Creation of the WasmEdge_VMContext.

WasmEdge_VMDelete⚠

Deletion of the WasmEdge_VMContext.

WasmEdge_VMExecute⚠

Invoke a WASM function by name.

WasmEdge_VMExecuteRegistered⚠

Invoke a WASM function by its module name and function name.

WasmEdge_VMGetActiveModule⚠

Get the current instantiated module in VM.

WasmEdge_VMGetExecutorContext⚠

Get the executor context used in the WasmEdge_VMContext.

WasmEdge_VMGetFunctionList⚠

Get the exported function list.

WasmEdge_VMGetFunctionListLength⚠

Get the length of exported function list.

WasmEdge_VMGetFunctionType⚠

Get the function type by function name.

WasmEdge_VMGetFunctionTypeRegistered⚠

Get the function type by function name.

WasmEdge_VMGetImportModuleContext⚠

Get the module instance corresponding to the WasmEdge_HostRegistration settings.

WasmEdge_VMGetLoaderContext⚠

Get the loader context used in the WasmEdge_VMContext.

WasmEdge_VMGetRegisteredModule⚠

Get the registered module in VM by the module name.

WasmEdge_VMGetStatisticsContext⚠

Get the statistics context used in the WasmEdge_VMContext.

WasmEdge_VMGetStoreContext⚠

Get the store context used in the WasmEdge_VMContext.

WasmEdge_VMGetValidatorContext⚠

Get the validator context used in the WasmEdge_VMContext.

WasmEdge_VMInstantiate⚠

Instantiate the validated WASM module in the VM context.

WasmEdge_VMListRegisteredModule⚠

List the registered module names in the WasmEdge_VMContext.

WasmEdge_VMListRegisteredModuleLength⚠

Get the length of registered module list in the WasmEdge_VMContext.

WasmEdge_VMLoadWasmFromASTModule⚠

Load the WASM module from loaded WasmEdge AST Module.

WasmEdge_VMLoadWasmFromBuffer⚠

Load the WASM module from a buffer.

WasmEdge_VMLoadWasmFromFile⚠

Load the WASM module from a WASM file.

WasmEdge_VMRegisterModuleFromASTModule⚠

Instantiate and register an AST Module into a named module instance in VM.

WasmEdge_VMRegisterModuleFromBuffer⚠

Register and instantiate WASM into the store in VM from a buffer.

WasmEdge_VMRegisterModuleFromFile⚠

Register and instantiate WASM into the store in VM from a WASM file.

WasmEdge_VMRegisterModuleFromImport⚠

Register a module instance into the store in VM with exporting its module name.

WasmEdge_VMRunWasmFromASTModule⚠

Instantiate the WASM module from a WasmEdge AST Module and invoke a function by name.

WasmEdge_VMRunWasmFromBuffer⚠

Instantiate the WASM module from a buffer and invoke a function by name.

WasmEdge_VMRunWasmFromFile⚠

Instantiate the WASM module from a WASM file and invoke a function by name.

WasmEdge_VMValidate⚠

Validate the WASM module loaded into the VM context.

WasmEdge_ValidatorCreate⚠

Creation of the WasmEdge_ValidatorContext.

WasmEdge_ValidatorDelete⚠

Deletion of the WasmEdge_ValidatorContext.

WasmEdge_ValidatorValidate⚠

Validate the WasmEdge AST Module.

WasmEdge_ValueGenExternRef⚠

Generate the function reference WASM value.

WasmEdge_ValueGenF32⚠

Generate the F32 WASM value.

WasmEdge_ValueGenF64⚠

Generate the F64 WASM value.

WasmEdge_ValueGenFuncRef⚠

Generate the function reference WASM value.

WasmEdge_ValueGenI32⚠

Generate the I32 WASM value.

WasmEdge_ValueGenI64⚠

Generate the I64 WASM value.

WasmEdge_ValueGenNullRef⚠

Generate the NULL reference WASM value.

WasmEdge_ValueGenV128⚠

Generate the V128 WASM value.

WasmEdge_ValueGetExternRef⚠

Retrieve the external reference from the WASM value.

WasmEdge_ValueGetF32⚠

Retrieve the F32 value from the WASM value.

WasmEdge_ValueGetF64⚠

Retrieve the F64 value from the WASM value.

WasmEdge_ValueGetFuncRef⚠

Retrieve the function instance context from the WASM value.

WasmEdge_ValueGetI32⚠

Retrieve the I32 value from the WASM value.

WasmEdge_ValueGetI64⚠

Retrieve the I64 value from the WASM value.

WasmEdge_ValueGetV128⚠

Retrieve the V128 value from the WASM value.

WasmEdge_ValueIsNullRef⚠

Specify the WASM value is a null reference or not.

WasmEdge_VersionGet⚠

Get the version string of the WasmEdge C API.

WasmEdge_VersionGetMajor⚠

Get the major version value of the WasmEdge C API.

WasmEdge_VersionGetMinor⚠

Get the minor version value of the WasmEdge C API.

WasmEdge_VersionGetPatch⚠

Get the patch version value of the WasmEdge C API.

Type Aliases

WasmEdge_CompilerOptimizationLevel

WasmEdge_CompilerOutputFormat

WasmEdge_ErrCategory

WasmEdge_ErrCode

WasmEdge_ExternalType

WasmEdge_HostFunc_t

WasmEdge_HostRegistration

WasmEdge_Mutability

WasmEdge_NumType

WasmEdge_ProgramOptionType

Type of option value.

WasmEdge_Proposal

WasmEdge_RefType

WasmEdge_ValType

WasmEdge_WrapFunc_t

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int128_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint128_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t