Function wasmedge_sys::ffi::WasmEdge_StringCreateByCString

pub unsafe extern "C" fn WasmEdge_StringCreateByCString(
    Str: *const c_char,
) -> WasmEdge_String

Creation of the WasmEdge_String with the C string.

The caller owns the object and should call WasmEdge_StringDelete to destroy it. This function only supports the C string with NULL termination. If the input string may have \0 character, please use the WasmEdge_StringCreateByBuffer instead.

\param Str the NULL-terminated C string to copy into the WasmEdge_String object.

\returns string object. Length will be 0 and Buf will be NULL if failed or the input string is a NULL.