wasmedge_sys::ffi
Function WasmEdge_VMAsyncRunWasmFromFile

pub unsafe extern "C" fn WasmEdge_VMAsyncRunWasmFromFile(
    Cxt: *mut WasmEdge_VMContext,
    Path: *const c_char,
    FuncName: WasmEdge_String,
    Params: *const WasmEdge_Value,
    ParamLen: u32,
) -> *mut WasmEdge_Async

Instantiate the WASM module from a WASM file and asynchronous invoke a function by name.

This is the function to invoke a WASM function rapidly. Load and instantiate the WASM module from the file path, and then invoke a function by name and parameters. If the Returns buffer length is smaller than the arity of the function, the overflowed return values will be discarded. After calling this function, a new anonymous module instance owned by VM is instantiated, and the old one will be destroyed.

The caller owns the object and should call WasmEdge_AsyncDelete to destroy it.

This function is thread-safe.

\param Cxt the WasmEdge_VMContext. \param Path the NULL-terminated C string of the WASM file path. \param FuncName the function name WasmEdge_String. \param Params the WasmEdge_Value buffer with the parameter values. \param ParamLen the parameter buffer length.

\returns WasmEdge_Async. Call WasmEdge_AsyncGet for the result, and call WasmEdge_AsyncDelete to destroy this object.