wasmedge_sys::instance
Module module

Defines WasmEdge Instance and other relevant types.

Structs

ImportModule

An ImportModule represents a host module with a name. A host module consists of one or more host function, table, memory, and global instances, which are defined outside wasm modules and fed into wasm modules as imports.

Instance

An Instance represents an instantiated module. In the instantiation process, An Instance is created from alModule. From an Instance the exported functions, tables, memories, and globals can be fetched.

WasiModule

A WasiModule is a module instance for the WASI specification.

Traits

AsInstance

The object as an module instance is required to implement this trait.