Module wasmedge_sys::instance::table

Defines WasmEdge Table and TableType structs.

A WasmEdge Table defines a WebAssembly table instance described by its TableType. TableType specifies the limits on the size of a table. The start of the limit range specifies the lower bound (inclusive) of the size, while the end resticts the upper bound (inclusive).

Structs

Table

A WasmEdge Table defines a WebAssembly table instance described by its type. A table is an array-like structure and stores function references.

TableType

A WasmEdge TableType classifies a Table instance over elements of element types within a size range.