Module wasmedge_sys::plugin

Defines PluginManager and Plugin structs.

Structs

ModuleDescriptor

Defines the module descriptor for plugins.

Plugin

Represents a loaded plugin. It provides the APIs for accessing the plugin.

PluginDescriptor

Represents Plugin descriptor for plugins.

PluginManager

Defines the APIs for loading plugins and check the basic information of the loaded plugins.

PluginVersion

Defines the version of a plugin.

ProgramOption

Defines the program option for plugins.

Enums

ProgramOptionType

Defines the type of the program options.

Type Aliases

ModuleInstanceCreateFn

Defines the type of the function that creates a module instance for a plugin.

PluginModule

Represents a Plugin module instance.