Struct wasmedge_sys::GlobalType

pub struct GlobalType { /* private fields */ }

Defines the type of a wasm global variable.

A GlobalType classifies a global variable that hold a value and can either be mutable or immutable.

Implementations

impl GlobalType

pub fn create(val_ty: ValType, mutable: Mutability) -> WasmEdgeResult<Self>

Create a new GlobalType to be associated with the given ValType and Mutability.

Arguments

val_type - The value type of the global variable.

mutability - The mutability of the global variable.

Errors

If fail to create a new GlobalType, then an error is returned.

pub fn value_type(&self) -> ValType

Returns the value type of the GlobalType.

pub fn mutability(&self) -> Mutability

Returns the Mutability value of the GlobalType.

pub fn as_ptr(&self) -> *const WasmEdge_GlobalTypeContext

Provides a raw pointer to the inner global type context.