wasmedge_sys
Struct Loader

pub struct Loader { /* private fields */ }

Loader is used to load WASM modules from the given WASM files or buffers.

Implementations

impl Loader

pub fn create(config: Option<&Config>) -> WasmEdgeResult<Self>

Create a new Loader to be associated with the given global configuration.

Arguments

config - A global configuration.

Error

If fail to create a Loader, then an error is returned.

pub fn from_file(&self, file: impl AsRef<Path>) -> WasmEdgeResult<Arc<Module>>

Loads a WASM module from a WASM file.

Arguments

file - A wasm file or an AOT wasm file.

Error

If fail to load, then an error is returned.

Example

let file = "path/to/foo.wasm"
let module = loader.from_file(file)?;

pub fn from_bytes(&self, bytes: impl AsRef<[u8]>) -> WasmEdgeResult<Arc<Module>>

Loads a WASM module from a in-memory bytes.

Arguments

bytes - A in-memory WASM bytes.

Error

If fail to load, then an error is returned.

Example

let bytes = b"\0asm\x01\0\0\0";
let module = loader.from_bytes(&bytes)?;

Note that the text format is not accepted:

assert!(loader.from_bytes(b"(module)").is_err());