Struct wasmedge_sys::Store

pub struct Store { /* private fields */ }

Struct of Wasmedge Store.

The Store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals that have been allocated during the life time of the Vm.

Implementations

impl Store

pub fn create() -> WasmEdgeResult<Self>

Creates a new Store.

Error

If fail to create, then an error is returned.

pub fn find_func(&self, name: impl AsRef<str>) -> WasmEdgeResult<Function>

Returns the exported function instance in the anonymous module by the given function name.

After instantiating a WASM module, the WASM module is registered into the Store as an anonymous module.

Argument

name specifies the target exported function instance.

Error

If fail to find the target function, then an error is returned.

pub fn find_func_registered(
&self,
mod_name: impl AsRef<str>,
func_name: impl AsRef<str>
) -> WasmEdgeResult<Function>

Returns the exported function instance in the registered module by the given function name and module name.

Arguments

mod_name specifies the name of the registered module.

func_name specifies the name of the exported function instance.

Error

If fail to find the target registered function, then an error is returned.

pub fn find_table(&self, name: impl AsRef<str>) -> WasmEdgeResult<Table>

Returns the exported table instance in the anonymous module by the given table name.

After instantiating a WASM module, the WASM module is registered into the Store as an anonymous module.

Argument

name specifies the target exported table instance.

Error

If fail to find the target table, then an error is returned.

pub fn find_table_registered(
&self,
mod_name: impl AsRef<str>,
table_name: impl AsRef<str>
) -> WasmEdgeResult<Table>

Returns the exported table instance in the registered module by the given table name and module name.

Arguments

mod_name specifies the name of the registered module.

table_name specifies the name of the exported table instance.

Error

If fail to find the target registered table, then an error is returned.

pub fn find_memory(&self, name: impl AsRef<str>) -> WasmEdgeResult<Memory>

Returns the exported memory instance in the anonymous module by the given memory name.

After instantiating a WASM module, the WASM module is registered into the Store as an anonymous module.

Argument

name specifies the target exported memory instance.

Error

If fail to find the target memory, then an error is returned.

pub fn find_memory_registered(
&self,
mod_name: impl AsRef<str>,
mem_name: impl AsRef<str>
) -> WasmEdgeResult<Memory>

Returns the exported memory instance in the registered module by the given memory name and module name.

Arguments

mod_name specifies the name of the registered module.

mem_name specifies the name of the exported memory instance.

Error

If fail to get the target registered memory, then an error is returned.

pub fn find_global(&self, name: impl AsRef<str>) -> WasmEdgeResult<Global>

Returns the exported global instance in the anonymous module by the given global name.

After instantiating a WASM module, the WASM module is registered into the Store as an anonymous module.

Argument

name specifies the target exported global instance.

Error

If fail to find the target global, then an error is returned.

pub fn find_global_registered(
&self,
mod_name: impl AsRef<str>,
global_name: impl AsRef<str>
) -> WasmEdgeResult<Global>

Returns the exported global instance in the registered module by the given global name and module name.

Arguments

mod_name specifies the name of the registered module.

global_name specifies the name of the exported global instance.

Error

If fail to find the target registered global, then an error is returned.

pub fn func_len(&self) -> u32

Returns the length of the exported functions in the anonymous module.

pub fn func_names(&self) -> Option<Vec<String>>

Returns the names of the exported functions in the anonymous module.

pub fn reg_func_len(&self, mod_name: impl AsRef<str>) -> u32

Returns the length of the exported functions in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn reg_func_names(&self, mod_name: impl AsRef<str>) -> Option<Vec<String>>

Returns the names of the exported functions in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn table_len(&self) -> u32

Returns the length of the exported tables in the anonymous module.

pub fn table_names(&self) -> Option<Vec<String>>

Returns the names of the exported tables in the anonynous module.

pub fn reg_table_len(&self, mod_name: impl AsRef<str>) -> u32

Returns the length of the exported tables in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn reg_table_names(&self, mod_name: impl AsRef<str>) -> Option<Vec<String>>

Returns the names of the exported tables in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn global_len(&self) -> u32

Returns the length of the exported globals in the anonymous module.

pub fn global_names(&self) -> Option<Vec<String>>

Returns the names of the exported globals in the anonymous module.

pub fn reg_global_len(&self, mod_name: impl AsRef<str>) -> u32

Returns the length of the exported globals in the reigstered module.

Argument

mod_name specifies the name of the registered module.

pub fn reg_global_names(&self, mod_name: impl AsRef<str>) -> Option<Vec<String>>

Returns the names of all exported globals in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn mem_len(&self) -> u32

Returns the length of the exported memories in the anonymous module.

pub fn mem_names(&self) -> Option<Vec<String>>

Returns the names of all exported memories in the anonymous module.

pub fn reg_mem_len(&self, mod_name: impl AsRef<str>) -> u32

Returns the length of the exported memories in the registered module.

Argument

mod_name specifies the name of the registered module.

pub fn reg_mem_names(&self, mod_name: impl AsRef<str>) -> Option<Vec<String>>

Returns the names of all exported memories in the registered module.

pub fn reg_module_len(&self) -> u32

Returns the length of the registered modules.

pub fn reg_module_names(&self) -> Option<Vec<String>>

Returns the names of all registered modules.

pub fn active_module(&mut self) -> WasmEdgeResult<Instance<'_>>

Returns the active anonymous module instance.

Error

If fail to find the target module instance, then an error is returned.

pub fn named_module(
&mut self,
name: impl AsRef<str>
) -> WasmEdgeResult<Instance<'_>>

Returns the registered module instance by the module name.

Argument

mod_name specifies the name of the registered module instance.

Error

If fail to find the target module instance, then an error is returned.

pub fn contains_func(&self, func_name: impl AsRef<str>) -> WasmEdgeResult<()>

Checks if the Store contains a function of which the name matches the given func_name.

Argument

func_name specifies the function’s name to check.

Error

If fail to find the name in the Store, then an error is returned.

pub fn contains_reg_func(
&self,
mod_name: impl AsRef<str>,
func_name: impl AsRef<str>
) -> WasmEdgeResult<()>

Checks if the Store contains a registered function of which the name matches the given func_name.

Argument

func_name specifies the registered function’s name to check.

Error

If fail to find the name in the Store, then an error is returned.

pub fn contains_mod_name(&self, mod_name: impl AsRef<str>) -> WasmEdgeResult<()>

Checks if the Store contains a registered module of which the name matches the given mod_name.

Argument

mod_name specifies the registered module’s name to check.

Error

If fail to find the name in the Store, then an error is returned.