Struct wasmedge_sys::Table

pub struct Table { /* private fields */ }

A WasmEdge Table defines a WebAssembly table instance described by its type. A table is an array-like structure and stores function references.

This example shows how to use Table to store and retrieve function references.

Implementations

impl Table

pub fn create(ty: &TableType) -> WasmEdgeResult<Self>

Creates a new Table to be associated with the given element type and the size.

Arguments

ty specifies the type of the new Table.

Error

If fail to create the table instance, then WasmEdgeError::Table(TableError::Create)(crate::error::TableError) is returned.

Example

use wasmedge_sys::{TableType, Table};
use wasmedge_types::RefType;

// create a TableType instance
let ty = TableType::create(RefType::FuncRef, 10, Some(20)).expect("fail to create a TableType");

// create a Table instance
let table = Table::create(&ty).expect("fail to create a Table");

pub fn ty(&self) -> WasmEdgeResult<TableType>

Returns the TableType of the Table.

Error

If fail to get type, then an error is returned.

pub fn get_data(&self, idx: u32) -> WasmEdgeResult<WasmValue>

Returns the element value at a specific position in the Table.

Arguments

idx specifies the position in the Table, at which the WasmValue is returned.

Error

If fail to get the data, then an error is returned.

pub fn set_data(&mut self, data: WasmValue, idx: u32) -> WasmEdgeResult<()>

Sets a new element value at a specific position in the Table.

Arguments

data specifies the new data.

idx specifies the position of the new data to be stored in the Table.

Error

If fail to set data, then an error is returned.

pub fn capacity(&self) -> usize

Returns the capacity of the Table.

Example

use wasmedge_sys::{TableType, Table};
use wasmedge_types::RefType;

// create a TableType instance and a Table
let ty = TableType::create(RefType::FuncRef, 10, Some(20)).expect("fail to create a TableType");
let table = Table::create(&ty).expect("fail to create a Table");

// check capacity
assert_eq!(table.capacity(), 10);

pub fn grow(&mut self, size: u32) -> WasmEdgeResult<()>

Increases the capacity of the Table.

After growing, the new capacity must be in the range defined by limit when the table is created.

Argument

size specifies the size to be added to the Table.

Error

If fail to increase the size of the Table, then an error is returned.

pub fn as_ptr(&self) -> *const WasmEdge_TableInstanceContext

Provides a raw pointer to the inner table context.