Struct wasmedge_sys::Vm

pub struct Vm { /* private fields */ }

A Vm defines a virtual environment for managing WebAssembly programs.

Implementations

impl Vm

pub fn create(
config: Option<Config>,
store: Option<&mut Store>
) -> WasmEdgeResult<Self>

Creates a new Vm to be associated with the given configuration and store.

Arguments

config - A configuration for the new Vm.

store - An external WASM store used by the new Vm. The instantiation and execution of the new Vm will consume this store context. If no store context is specified when creating a Vm, then the Vm itself will allocate and own a store.

Error

If fail to create, then an error is returned.

pub fn register_wasm_from_file(
&self,
mod_name: impl AsRef<str>,
file: impl AsRef<Path>
) -> WasmEdgeResult<()>

Registers and instantiates a WASM module from a wasm file into the Vm.

The workflow of the function can be summarized as the following steps:

First, loads a WASM module from a given path, then

Registers all exported instances in the WASM module into the Vm;

Finally, instantiates the exported instances.

Arguments

mod_name - The name for the WASM module to be registered.

file - The wasm file, of which the file extension should be one of wasm, wat, dylib on macOS, so on Linux or dll on Windows.

Error

If fail to register the target WASM, then an error is returned.

pub fn register_wasm_from_import(
&mut self,
import: ImportObject
) -> WasmEdgeResult<()>

Registers a given WasmEdge module instance that implements the ImportObject trait into the Vm, and instantiates it.

The workflow of the function can be summarized as the following steps:

First, registers the exported module instance into the Vm, then

Instatiates the exported instances.

Argument

import - The module instance to be registered.

Error

If fail to register the WASM module, then an error is returned.

pub fn register_wasm_from_bytes(
&self,
mod_name: impl AsRef<str>,
bytes: &[u8]
) -> WasmEdgeResult<()>

Registers a WASM module from a given in-memory wasm bytes into the Vm, and instantiates it.

The workflow of the function can be summarized as the following steps:

First, loads a WASM module from the given in-memory wasm bytes, then

Registers all exported instances in the WASM module into the Vm;

Finally, instantiates the exported instances.

Arguments

mod_name - The name of the WASM module to be registered.

bytes - The in-memory wasm bytes.

Error

If fail to register the WASM module, then an error is returned.

pub fn register_wasm_from_module(
&self,
mod_name: impl AsRef<str>,
module: Module
) -> WasmEdgeResult<()>

Registers a given WasmEdge AST Module into the Vm, and instantiates it. The module is consumed.

The workflow of the function can be summarized as the following steps:

First, loads a WASM module from the given WasmEdge AST Module, then

Registers all exported instances in the WASM module into the Vm;

Finally, instantiates the exported instances.

Arguments

mod_name - The name of the WASM module to be registered.

module - The WasmEdge AST Module generated by Loader or Compiler.

Error

If fail to register the WASM module, then an error is returned.

pub fn run_wasm_from_file(
&self,
file: impl AsRef<Path>,
func_name: impl AsRef<str>,
params: impl IntoIterator<Item = WasmValue>
) -> WasmEdgeResult<Vec<WasmValue>>

Runs a function defined in a WASM file.

The workflow of the function can be summarized as the following steps:

First, loads the WASM module from a given WASM file, and validates the loaded module; then,

Instantiates the WASM module; finally,

Invokes a function by name and parameters.

Arguments

file - The wasm file, of which the file extension should be one of wasm, wat, dylib on macOS, so on Linux or dll on Windows.

func_name - The name of the function.

params - The the parameter values which are used by the function.

Error

If fail to run, then an error is returned.

pub fn run_wasm_from_bytes(
&self,
bytes: &[u8],
func_name: impl AsRef<str>,
params: impl IntoIterator<Item = WasmValue>
) -> WasmEdgeResult<Vec<WasmValue>>

Runs a function from a in-memory wasm bytes.

The workflow of the function can be summarized as the following steps:

First, loads and instantiates the WASM module from a buffer, then

Invokes a function by name and parameters.

Arguments

bytes - The in-memory wasm bytes.

func_name - The name of the function.

params - The parameter values which are used by the function.

Error

If fail to run, then an error is returned.

pub fn run_wasm_from_module(
&self,
module: Module,
func_name: impl AsRef<str>,
params: impl IntoIterator<Item = WasmValue>
) -> WasmEdgeResult<Vec<WasmValue>>

Runs a function from a WasmEdge compiled Module.

The workflow of the function can be summarized as the following steps:

First, loads and instantiates the WASM module from a WasmEdge AST Module, then

Invokes a function by name and parameters.

Arguments

module - The WasmEdge AST Module generated by Loader or Compiler.

func_name - The name of the function.

params - The parameter values which are used by the function.

Error

If fail to run, then an error is returned.

pub fn load_wasm_from_module(&self, module: &Module) -> WasmEdgeResult<()>

Loads a WASM module from a WasmEdge AST Module.

This is the first step to invoke a WASM function step by step.

Argument

module - A WasmEdge AST Module generated by Loader or Compiler.

Error

If fail to load, then an error is returned.

pub fn load_wasm_from_bytes(&self, bytes: &[u8]) -> WasmEdgeResult<()>

Loads a WASM module from a in-memory WASM bytes. The loaded module is not validated.

This is the first step to invoke a WASM function step by step.

Argument

bytes - The in-memory wasm bytes.

Error

If fail to load, then an error is returned. The loaded module is not validated.

pub fn load_wasm_from_file(&self, file: impl AsRef<Path>) -> WasmEdgeResult<()>

Loads a WASM module from a WASM file.

This is the first step to invoke a WASM function step by step.

Argument

file - The wasm file, of which the file extension should be one of wasm, wat, dylib on macOS, so on Linux or dll on Windows.

Error

If fail to load, then an error is returned. The loaded module is not validated.

pub fn validate(&self) -> WasmEdgeResult<()>

Validates a WASM module loaded into the Vm.

This is the second step to invoke a WASM function step by step. After loading a WASM module into the Vm, call this function to validate it. Note that only validated WASM modules can be instantiated in the Vm.

Error

If fail to validate, then an error is returned.

pub fn instantiate(&self) -> WasmEdgeResult<()>

Instantiates a validated WASM module in the Vm.

This is the third step to invoke a WASM function step by step. After validating a WASM module in the Vm, call this function to instantiate it; then, call execute to invoke the exported function in this WASM module.

Error

If fail to instantiate, then an error is returned.

pub fn run_function(
&self,
func_name: impl AsRef<str>,
params: impl IntoIterator<Item = WasmValue>
) -> WasmEdgeResult<Vec<WasmValue>>

Runs an exported WASM function by name. The WASM function is hosted by the anonymous module in the store of the Vm.

This is the final step to invoke a WASM function step by step. After instantiating a WASM module in the Vm, the WASM module is registered into the store of the Vm as an anonymous module. Then repeatedly call this function to invoke the exported WASM functions by their names until the Vm is reset or a new WASM module is registered or loaded.

Arguments

func_name - The name of the exported WASM function to run.

params - The parameter values passed to the exported WASM function.

Error

If fail to run the WASM function, then an error is returned.

pub fn run_registered_function(
&self,
mod_name: impl AsRef<str>,
func_name: impl AsRef<str>,
params: impl IntoIterator<Item = WasmValue>
) -> WasmEdgeResult<Vec<WasmValue>>

Runs an exported WASM function by its name and the module’s name in which the WASM function is hosted.

After registering a WASM module in the Vm, repeatedly call this function to run exported WASM functions by their function names and the module names until the Vm is reset.

Arguments

mod_name - The name of the WASM module registered into the store of the Vm.

func_name - The name of the exported WASM function to run.

params - The parameter values passed to the exported WASM function.

Error

If fail to run the WASM function, then an error is returned.

pub fn get_function_type(
&self,
func_name: impl AsRef<str>
) -> WasmEdgeResult<FuncType>

Returns the function type of a WASM function by its name. The function is hosted in the anonymous module of the Vm.

Argument

func_name - The name of the target WASM function.

Error

If fail to get the function type, then an error is returned.

pub fn get_registered_function_type(
&self,
mod_name: impl AsRef<str>,
func_name: impl AsRef<str>
) -> WasmEdgeResult<FuncType>

Returns the function type of a WASM function by its name and the name of the registered module which hosts the WASM function.

Arguments

mod_name - The name of the registered module.

func_name - The name of the target WASM function.

pub fn reset(&mut self)

Resets the Vm.

pub fn function_list_len(&self) -> usize

Returns the number of the exported functions.

pub fn function_iter(
&self
) -> impl Iterator<Item = (Option<String>, Option<FuncType>)>

Returns an iterator of pairs of the exported function’s name and type.

pub fn wasi_module_mut(&mut self) -> WasmEdgeResult<WasiModule>

Returns the mutable Wasi module instance.

Notice that this function is only available when a config with the enabled wasi option is used in the creation of this Vm.

pub fn wasmedge_process_module_mut(
&mut self
) -> WasmEdgeResult<WasmEdgeProcessModule>

Returns the mutable WasmEdgeProcess module instance.

Notice that this function is only available when a config with the enabled wasmedge_process option is used in the creation of this Vm.

pub fn store_mut(&self) -> WasmEdgeResult<Store>

Returns the internal Store from the Vm.

pub fn statistics_mut(&self) -> WasmEdgeResult<Statistics>

Returns the internal Statistics from the Vm.

pub fn active_module(&self) -> WasmEdgeResult<Instance>

Returns the active (also called anonymous) module instance.

pub fn contains_module(&self, mod_name: impl AsRef<str>) -> bool

Checks if the Vm contains a registered module of which the name matches the given mod_name.

Argument

mod_name - The registered module’s name to check.

pub fn loader(&self) -> WasmEdgeResult<Loader>

Returns the internal Loader from the Vm.

pub fn validator(&self) -> WasmEdgeResult<Validator>

Returns the internal Validator from the Vm.

pub fn executor(&self) -> WasmEdgeResult<Executor>

Returns the internal Executor from the Vm.