Struct wasmedge_sys::WasiCryptoCommonModule

pub struct WasiCryptoCommonModule { /* private fields */ }

A WasiCryptoCommonModule is a module instance for the WASI-Crypto specification, covering common types and functions for symmetric operations.

Implementations

impl WasiCryptoCommonModule

pub fn create() -> WasmEdgeResult<Self>

Creates and initializes a wasi_crypto common host module named wasi_ephemeral_crypto_common, which contains the wasi_crypto common host functions.

Error

If the wasi_crypto plugin is not found, then an error is returned.