Struct wasmedge_sys::WasiModule

pub struct WasiModule { /* private fields */ }

A WasiModule is a module instance for the WASI specification.

Usage

WasiModule implements AsImport trait, therefore it can be used to register function, table, memory and global instances.

Example

A WasiModule can be created implicitly inside a Vm by passing the Vm a config argument in which the wasi option is enabled.

Example

Implementations

impl WasiModule

pub fn create(
args: Option<Vec<&str>>,
envs: Option<Vec<&str>>,
preopens: Option<Vec<&str>>
) -> WasmEdgeResult<Self>

Creates a WASI host module which contains the WASI host functions, and initializes it with the given parameters.

Arguments

args - The commandline arguments. The first argument is the program name.

envs - The environment variables in the format ENV_VAR_NAME=VALUE.

preopens - The directories to pre-open. The required format is DIR1:DIR2.

Error

If fail to create a host module, then an error is returned.

pub fn init_wasi(
&mut self,
args: Option<Vec<&str>>,
envs: Option<Vec<&str>>,
preopens: Option<Vec<&str>>
)

Initializes the WASI host module with the given parameters.

Arguments

args - The commandline arguments. The first argument is the program name.

envs - The environment variables in the format ENV_VAR_NAME=VALUE.

preopens - The directories to pre-open. The required format is DIR1:DIR2.

pub fn exit_code(&self) -> u32

Returns the WASI exit code.

The WASI exit code can be accessed after running the “_start” function of a wasm32-wasi program.

pub fn get_native_handler(&self, fd: i32) -> WasmEdgeResult<u64>

Returns the native handler from the mapped FD/Handler.

Argument

fd - The WASI mapped Fd.

Error

If fail to get the native handler, then an error is returned.