Trait wasmedge_sys::ImportInstance

pub trait ImportInstance {
    fn add_func(&mut self, name: impl AsRef<str>, func: Function);
    fn add_table(&mut self, name: impl AsRef<str>, table: Table);
    fn add_memory(&mut self, name: impl AsRef<str>, memory: Memory);
    fn add_global(&mut self, name: impl AsRef<str>, global: Global);
}

The object to be registered into a Vm or an Executor instance is required to implement this trait. The object that implements this trait can be registered via the Vm::register_wasm_from_import function, or the Executor::register_import_object function.

Required Methods

fn add_func(&mut self, name: impl AsRef<str>, func: Function)

Imports a host function instance.

Arguments

name - The name of the host function instance to import.

func - The host function instance to import.

fn add_table(&mut self, name: impl AsRef<str>, table: Table)

Imports a table instance.

Arguments

name - The name of the host table instance to import.

table - The host table instance to import.

fn add_memory(&mut self, name: impl AsRef<str>, memory: Memory)

Imports a memory instance.

Arguments

name - The name of the host memory instance to import.

memory - The host memory instance to import.

fn add_global(&mut self, name: impl AsRef<str>, global: Global)

Imports a global instance.

Arguments

name - The name of the host global instance to import.

global - The host global instance to import.