Crate wasmedge_sys

Rust WasmEdge bindings

This library contains safe Rust bindings for WasmEdge, a lightweight, high-performance, and extensible WebAssembly runtime for cloud native, edge, and decentralized applications.

WasmEdge 0.9.0 is the lowest supported version for the underlying library.

Most of this documentation is generated from the C API. Until all parts of the documentation have been reviewed there will be incongruities with the actual Rust API.

See also

WasmEdge Runtime

WasmEdge C API Documentation

Modules

error

Defines the error types.

instance

Defines WasmEdge instance structs, including Function, Global, Memory, and Table.

module

Defines WasmEdge AST Module, Export, and Import structs.

types

Defines the WebAssembly primitive types.

utils

Defines the versioning and logging functions.

Structs

Compiler

Struct of WasmEdge ahead-of-time(AOT) compiler.

Config

Struct of WasmEdge Config.

Executor

Struct of WasmEdge Executor.

Export

Struct of WasmEdge Export.

FuncType

Struct of WasmEdge FuncType.

Function

Struct of WasmEdge Function.

Global

Struct of WasmEdge Global.

GlobalType

Struct of WasmEdge GlobalType.

Import

Struct of WasmEdge Import.

ImportObject

Struct of WasmEdge ImportObject.

Instance

Struct of WasmEdge Instance.

Loader

Struct of WasmEdge Loader.

MemType

Struct of WasmEdge MemType.

Memory

Struct of WasmEdge Memory.

Module

Struct of WasmEdge Module.

Statistics

Struct of WasmEdge Statistics.

Store

Struct of Wasmedge Store.

Table

Struct of WasmEdge Table.

TableType

Struct of WasmEdge TableType

Validator

Struct of WasmEdge Validator.

Vm

Struct of WasmEdge Vm.

WasmValue

Struct of WasmEdge Value.

Enums

CompilerOptimizationLevel

Defines WasmEdge AOT compiler optimization level.

CompilerOutputFormat

Defines WasmEdge AOT compiler output binary format.

ExternalType

Defines WasmEdge ExternType values.

Mutability

Defines WasmEdge mutability values.

RefType

Defines reference types.

ValType

Defines value types.

Type Definitions

HostFunc

Type alias for a host function.

WasmEdgeResult

The WasmEdge result type.