Crate wasmedge_sys

Overview

The wasmedge-sys crate defines a group of low-level Rust APIs for WasmEdge, a light-weight, high-performance, and extensible WebAssembly runtime for cloud-native, edge, and decentralized applications.

For developers, it is strongly recommended that the APIs in wasmedge-sys are used to construct high-level libraries, while wasmedge-sdk is for building up business applications.

Notice that wasmedge-sdk requires Rust v1.63 or above in the stable channel.

Versioning Table

The following table provides the versioning information about each crate of WasmEdge Rust bindings.

wasmedge-sdk	WasmEdge lib	wasmedge-sys	wasmedge-types	wasmedge-macro

0.7.1	0.11.2	0.12.2	0.3.1	0.3.0

0.7.0	0.11.2	0.12	0.3.1	0.3.0

0.6.0	0.11.2	0.11	0.3.0	0.2.0

0.5.0	0.11.1	0.10	0.3.0	0.1.0

0.4.0	0.11.0	0.9	0.2.1	-

0.3.0	0.10.1	0.8	0.2	-

0.1.0	0.10.0	0.7	0.1	-

Build

To use or build the wasmedge-sys crate, the WasmEdge library is required.

If you choose to use install.sh to install WasmEdge Runtime on your local system. Please use WASMEDGE_INCLUDE_DIR and WASMEDGE_LIB_DIR to specify the paths to the include and lib directories, respectively. For example, use the following commands to specify the paths after using bash install.sh --path=$HOME/wasmedge-install to install WasmEdge Runtime on Ubuntu 20.04:

export WASMEDGE_INCLUDE_DIR=$HOME/wasmedge-install/include
export WASMEDGE_LIB_DIR=$HOME/wasmedge-install/lib

If you choose to manually download WasmEdge Runtime binary from WasmEdge Releases Page, it is strongly recommended to place it in $HOME/.wasmedge directory. It looks like below on Ubuntu 20.04. wasmedge-sys will search the directory automatically, you do not have to set any environment variables for it.

// $HOME/.wasmedge/
.
|-- bin
|   |-- wasmedge
|   `-- wasmedgec
|-- include
|   `-- wasmedge
|       |-- enum.inc
|       |-- enum_configure.h
|       |-- enum_errcode.h
|       |-- enum_types.h
|       |-- int128.h
|       |-- version.h
|       `-- wasmedge.h
`-- lib64
    |-- libwasmedge_c.so
    `-- wasmedge
        `-- libwasmedgePluginWasmEdgeProcess.so

5 directories, 11 files

Enable WasmEdge Plugins

If you’d like to enable WasmEdge Plugins (currently, only available on Linux platform), please use WASMEDGE_PLUGIN_PATH environment variable to specify the path to the directory containing the plugins. For example, use the following commands to specify the path on Ubuntu 20.04:

export WASMEDGE_PLUGIN_PATH=$HOME/.wasmedge/lib/wasmedge

Example

The following code presents how to use the APIs in wasmedge-sys to run a WebAssembly module written with its WAT format (textual format):

use wasmedge_sys::{Vm, WasmValue};
use wasmedge_types::wat2wasm;

fn main() -> Result<(), Box<dyn std::error::Error>> {
    // read the wasm bytes
    let wasm_bytes = wat2wasm(
        br#"
        (module
            (export "fib" (func $fib))
            (func $fib (param $n i32) (result i32)
             (if
              (i32.lt_s
               (get_local $n)
               (i32.const 2)
              )
              (return
               (i32.const 1)
              )
             )
             (return
              (i32.add
               (call $fib
                (i32.sub
                 (get_local $n)
                 (i32.const 2)
                )
               )
               (call $fib
                (i32.sub
                 (get_local $n)
                 (i32.const 1)
                )
               )
              )
             )
            )
           )
"#,
    )?;

    // create a Vm instance
    let mut vm = Vm::create(None, None)?;

    // register the wasm bytes
    let module_name = "extern-module";
    vm.register_wasm_from_bytes(module_name, &wasm_bytes)?;

    // run the exported function named "fib"
    let func_name = "fib";
    let result = vm.run_registered_function(module_name, func_name, [WasmValue::from_i32(5)])?;

    assert_eq!(result.len(), 1);
    assert_eq!(result[0].to_i32(), 8);

    Ok(())
}

[Click for more examples]

See also

WasmEdge Runtime Official Website

WasmEdge Docs

WasmEdge C API Documentation

Modules

frame

Defines WasmEdge CallingFrame.

types

Defines the WebAssembly primitive types.

utils

Defines the versioning and logging functions.

Structs

CallingFrame

Represents a calling frame on top of stack.

Config

Defines Config struct used to check/set the configuration options.

Executor

Defines an execution environment for both pure WASM and compiled WASM.

ExportType

Defines the types of the exported wasm values.

FuncRef

Defines a reference to a host function.

FuncType

Defines the type of a host function.

Function

Defines a host function.

Global

Defines a WebAssembly global variable, which stores a single value of the given type and a flag indicating whether it is mutable or not.

GlobalType

Defines the type of a wasm global variable.

ImportModule

An ImportModule represents a host module with a name. A host module consists of one or more host function, table, memory, and global instances, which are defined outside wasm modules and fed into wasm modules as imports.

ImportType

Defines the types of the imported wasm value.

Instance

An Instance represents an instantiated module. In the instantiation process, An Instance is created from alModule. From an Instance the exported functions, tables, memories, and globals can be fetched.

Loader

Loader is used to load WASM modules from the given WASM files or buffers.

MemType

Defines the type of a wasm memory instance

Memory

Defines a WebAssembly memory instance, which is a linear memory described by its type. Each memory instance consists of a vector of bytes and an optional maximum size, and its size is a multiple of the WebAssembly page size (64KiB of each page).

Module

Defines compiled in-memory representation of an input WASM binary.

Statistics

Struct of WasmEdge Statistics.

Store

A Store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals that have been allocated during the life time of the Vm.

Table

A WasmEdge Table defines a WebAssembly table instance described by its type. A table is an array-like structure and stores function references.

TableType

A WasmEdge TableType classifies a Table instance over elements of element types within a size range.

Validator

Struct of WasmEdge Validator.

Vm

A Vm defines a virtual environment for managing WebAssembly programs.

WasiModule

A WasiModule is a module instance for the WASI specification.

WasmEdgeProcessModule

A WasmEdgeProcessModule is a module instance for the WasmEdge_Process specification.

WasmValue

Defines a WebAssembly value.

Enums

ImportObject

Defines three types of module instances that can be imported into a WasmEdge Store instance.

Traits

AsImport

The object to be registered into a Vm or an Executor instance is required to implement this trait. The object that implements this trait can be registered via the Vm::register_wasm_from_import function, or the Executor::register_import_object function.

AsInstance

The object as an module instance is required to implement this trait.

Engine

The object that is used to perform a host function is required to implement this trait.

Type Definitions

BoxedFn

Type alias for a boxed native function. This type is used in thread-safe cases.