wasmedge_sys
Module ffi

Foreign function interfaces generated from WasmEdge C-API.

Structs

WasmEdge_ASTModuleContext

WasmEdge_Async

WasmEdge_Bytes

WasmEdge bytes struct.

WasmEdge_CallingFrameContext

WasmEdge_CompilerContext

WasmEdge_ConfigureContext

WasmEdge_ExecutorContext

WasmEdge_ExportTypeContext

WasmEdge_FunctionInstanceContext

WasmEdge_FunctionTypeContext

WasmEdge_GlobalInstanceContext

WasmEdge_GlobalTypeContext

WasmEdge_ImportTypeContext

WasmEdge_Limit

Struct of WASM limit.

WasmEdge_LoaderContext

WasmEdge_MemoryInstanceContext

WasmEdge_MemoryTypeContext

WasmEdge_ModuleDescriptor

Module descriptor for plugins.

WasmEdge_ModuleInstanceContext

WasmEdge_PluginContext

WasmEdge_PluginDescriptor

Plugin descriptor for plugins.

WasmEdge_PluginVersionData

Version data for plugins.

WasmEdge_ProgramOption

Program option for plugins.

WasmEdge_Result

WasmEdge result struct.

WasmEdge_StatisticsContext

WasmEdge_StoreContext

WasmEdge_String

WasmEdge string struct.

WasmEdge_TableInstanceContext

WasmEdge_TableTypeContext

WasmEdge_TagInstanceContext

WasmEdge_TagTypeContext

WasmEdge_VMContext

WasmEdge_ValType

WasmEdge WASM value type struct.

WasmEdge_ValidatorContext

WasmEdge_Value

WasmEdge WASM value struct.

Constants

WasmEdge_CompilerOptimizationLevel_O0

WasmEdge_CompilerOptimizationLevel_O1

WasmEdge_CompilerOptimizationLevel_O2

WasmEdge_CompilerOptimizationLevel_O3

WasmEdge_CompilerOptimizationLevel_Os

WasmEdge_CompilerOptimizationLevel_Oz

WasmEdge_CompilerOutputFormat_Native

WasmEdge_CompilerOutputFormat_Wasm

WasmEdge_ErrCategory_UserLevelError

WasmEdge_ErrCategory_WASM

WasmEdge_ErrCode_AOTDisabled

WasmEdge_ErrCode_AccessNullArray

WasmEdge_ErrCode_AccessNullException

WasmEdge_ErrCode_AccessNullFunc

WasmEdge_ErrCode_AccessNullI31

WasmEdge_ErrCode_AccessNullStruct

WasmEdge_ErrCode_ArrayOutOfBounds

WasmEdge_ErrCode_ArrayTypesMismatch

WasmEdge_ErrCode_ArrayTypesNumtypeRequired

WasmEdge_ErrCode_CastFailed

WasmEdge_ErrCode_CastNullToNonNull

WasmEdge_ErrCode_ConstExprRequired

WasmEdge_ErrCode_CoreInvalidExport

WasmEdge_ErrCode_CostLimitExceeded

WasmEdge_ErrCode_DataCountRequired

WasmEdge_ErrCode_DataSegDoesNotFit

WasmEdge_ErrCode_DivideByZero

WasmEdge_ErrCode_DupExportName

WasmEdge_ErrCode_ENDCodeExpected

WasmEdge_ErrCode_ElemSegDoesNotFit

WasmEdge_ErrCode_ExpectSharedMemory

WasmEdge_ErrCode_ExpectedZeroByte

WasmEdge_ErrCode_FuncNotFound

WasmEdge_ErrCode_FuncSigMismatch

WasmEdge_ErrCode_HostFuncError

WasmEdge_ErrCode_IllegalGrammar

WasmEdge_ErrCode_IllegalOpCode

WasmEdge_ErrCode_IllegalPath

WasmEdge_ErrCode_ImmutableArray

WasmEdge_ErrCode_ImmutableField

WasmEdge_ErrCode_ImmutableGlobal

WasmEdge_ErrCode_IncompatibleDataCount

WasmEdge_ErrCode_IncompatibleFuncCode

WasmEdge_ErrCode_IncompatibleImportType

WasmEdge_ErrCode_IndirectCallTypeMismatch

WasmEdge_ErrCode_InstrTypeMismatch

WasmEdge_ErrCode_IntegerOverflow

WasmEdge_ErrCode_IntegerTooLarge

WasmEdge_ErrCode_IntegerTooLong

WasmEdge_ErrCode_Interrupted

WasmEdge_ErrCode_IntrinsicsTableNotFound

WasmEdge_ErrCode_InvalidAlignment

WasmEdge_ErrCode_InvalidBrRefType

WasmEdge_ErrCode_InvalidCanonOption

WasmEdge_ErrCode_InvalidConvToInt

WasmEdge_ErrCode_InvalidCoreSort

WasmEdge_ErrCode_InvalidDataIdx

WasmEdge_ErrCode_InvalidElemIdx

WasmEdge_ErrCode_InvalidFieldIdx

WasmEdge_ErrCode_InvalidFuncIdx

WasmEdge_ErrCode_InvalidFuncTypeIdx

WasmEdge_ErrCode_InvalidGlobalIdx

WasmEdge_ErrCode_InvalidLabelIdx

WasmEdge_ErrCode_InvalidLaneIdx

WasmEdge_ErrCode_InvalidLimit

WasmEdge_ErrCode_InvalidLocalIdx

WasmEdge_ErrCode_InvalidMemPages

WasmEdge_ErrCode_InvalidMemoryIdx

WasmEdge_ErrCode_InvalidMut

WasmEdge_ErrCode_InvalidNotDefaultableArray

WasmEdge_ErrCode_InvalidNotDefaultableField

WasmEdge_ErrCode_InvalidPackedArray

WasmEdge_ErrCode_InvalidPackedField

WasmEdge_ErrCode_InvalidRefIdx

WasmEdge_ErrCode_InvalidResultArity

WasmEdge_ErrCode_InvalidStartFunc

WasmEdge_ErrCode_InvalidStoreAlignment

WasmEdge_ErrCode_InvalidSubType

WasmEdge_ErrCode_InvalidTableIdx

WasmEdge_ErrCode_InvalidTagIdx

WasmEdge_ErrCode_InvalidTagResultType

WasmEdge_ErrCode_InvalidUninitLocal

WasmEdge_ErrCode_InvalidUnpackedArray

WasmEdge_ErrCode_InvalidUnpackedField

WasmEdge_ErrCode_JunkSection

WasmEdge_ErrCode_LengthOutOfBounds

WasmEdge_ErrCode_MalformedAliasTarget

WasmEdge_ErrCode_MalformedCanonical

WasmEdge_ErrCode_MalformedCoreInstance

WasmEdge_ErrCode_MalformedDefType

WasmEdge_ErrCode_MalformedElemType

WasmEdge_ErrCode_MalformedExportKind

WasmEdge_ErrCode_MalformedFlagsType

WasmEdge_ErrCode_MalformedImportKind

WasmEdge_ErrCode_MalformedInstance

WasmEdge_ErrCode_MalformedMagic

WasmEdge_ErrCode_MalformedName

WasmEdge_ErrCode_MalformedRecordType

WasmEdge_ErrCode_MalformedRefType

WasmEdge_ErrCode_MalformedSection

WasmEdge_ErrCode_MalformedSort

WasmEdge_ErrCode_MalformedTable

WasmEdge_ErrCode_MalformedTupleType

WasmEdge_ErrCode_MalformedUTF8

WasmEdge_ErrCode_MalformedValType

WasmEdge_ErrCode_MalformedVariantType

WasmEdge_ErrCode_MalformedVersion

WasmEdge_ErrCode_MemoryOutOfBounds

WasmEdge_ErrCode_ModuleNameConflict

WasmEdge_ErrCode_MultiMemories

WasmEdge_ErrCode_MultiTables

WasmEdge_ErrCode_NonNullRequired

WasmEdge_ErrCode_NotValidated

WasmEdge_ErrCode_ReadError

WasmEdge_ErrCode_RefTypeMismatch

WasmEdge_ErrCode_RuntimeError

WasmEdge_ErrCode_SectionSizeMismatch

WasmEdge_ErrCode_SetValueErrorType

WasmEdge_ErrCode_SetValueToConst

WasmEdge_ErrCode_SharedMemoryNoMax

WasmEdge_ErrCode_Success

WasmEdge_ErrCode_TableOutOfBounds

WasmEdge_ErrCode_Terminated

WasmEdge_ErrCode_TooManyLocals

WasmEdge_ErrCode_TypeCheckFailed

WasmEdge_ErrCode_UnalignedAtomicAccess

WasmEdge_ErrCode_UncaughtException

WasmEdge_ErrCode_UndefinedElement

WasmEdge_ErrCode_UnexpectedEnd

WasmEdge_ErrCode_UninitializedElement

WasmEdge_ErrCode_UnknownCanonicalOption

WasmEdge_ErrCode_UnknownImport

WasmEdge_ErrCode_Unreachable

WasmEdge_ErrCode_UserDefError

WasmEdge_ErrCode_WrongInstanceAddress

WasmEdge_ErrCode_WrongInstanceIndex

WasmEdge_ErrCode_WrongVMWorkflow

WasmEdge_ExternalType_Function

WasmEdge_ExternalType_Global

WasmEdge_ExternalType_Memory

WasmEdge_ExternalType_Table

WasmEdge_ExternalType_Tag

WasmEdge_HostRegistration_Wasi

WasmEdge_Mutability_Const

WasmEdge_Mutability_Var

WasmEdge_Plugin_CurrentAPIVersion

WasmEdge_ProgramOptionType_Double

Boolean value.

WasmEdge_ProgramOptionType_Float

Boolean value.

WasmEdge_ProgramOptionType_Int8

Boolean value.

WasmEdge_ProgramOptionType_Int16

Boolean value.

WasmEdge_ProgramOptionType_Int32

Boolean value.

WasmEdge_ProgramOptionType_Int64

Boolean value.

WasmEdge_ProgramOptionType_None

No option value.

WasmEdge_ProgramOptionType_String

WasmEdge_String.

WasmEdge_ProgramOptionType_Toggle

Boolean value.

WasmEdge_ProgramOptionType_UInt8

Boolean value.

WasmEdge_ProgramOptionType_UInt16

Boolean value.

WasmEdge_ProgramOptionType_UInt32

Boolean value.

WasmEdge_ProgramOptionType_UInt64

Boolean value.

WasmEdge_Proposal_Annotations

WasmEdge_Proposal_BulkMemoryOperations

WasmEdge_Proposal_Component

WasmEdge_Proposal_ExceptionHandling

WasmEdge_Proposal_ExtendedConst

WasmEdge_Proposal_FunctionReferences

WasmEdge_Proposal_GC

WasmEdge_Proposal_ImportExportMutGlobals

WasmEdge_Proposal_Memory64

WasmEdge_Proposal_MultiMemories

WasmEdge_Proposal_MultiValue

WasmEdge_Proposal_NonTrapFloatToIntConversions

WasmEdge_Proposal_ReferenceTypes

WasmEdge_Proposal_RelaxSIMD

WasmEdge_Proposal_SIMD

WasmEdge_Proposal_SignExtensionOperators

WasmEdge_Proposal_TailCall

WasmEdge_Proposal_Threads

WasmEdge_TypeCode_AnyRef

WasmEdge_TypeCode_Array

WasmEdge_TypeCode_ArrayRef

WasmEdge_TypeCode_Epsilon

WasmEdge_TypeCode_EqRef

WasmEdge_TypeCode_ExnRef

WasmEdge_TypeCode_ExternRef

WasmEdge_TypeCode_F32

WasmEdge_TypeCode_F64

WasmEdge_TypeCode_Func

WasmEdge_TypeCode_FuncRef

WasmEdge_TypeCode_I8

WasmEdge_TypeCode_I16

WasmEdge_TypeCode_I32

WasmEdge_TypeCode_I64

WasmEdge_TypeCode_I31Ref

WasmEdge_TypeCode_NullExternRef

WasmEdge_TypeCode_NullFuncRef

WasmEdge_TypeCode_NullRef

WasmEdge_TypeCode_Rec

WasmEdge_TypeCode_Ref

WasmEdge_TypeCode_RefNull

WasmEdge_TypeCode_String

WasmEdge_TypeCode_Struct

WasmEdge_TypeCode_StructRef

WasmEdge_TypeCode_Sub

WasmEdge_TypeCode_SubFinal

WasmEdge_TypeCode_TypeIndex

WasmEdge_TypeCode_V128

Functions

WasmEdge_ASTModuleDeleteโ

Deletion of the WasmEdge_ASTModuleContext.

WasmEdge_ASTModuleListExportsโ

List the exports of the AST module.

WasmEdge_ASTModuleListExportsLengthโ

Get the length of exports list of the AST module.

WasmEdge_ASTModuleListImportsโ

List the imports of the AST module.

WasmEdge_ASTModuleListImportsLengthโ

Get the length of imports list of the AST module.

WasmEdge_AsyncCancelโ

Cancel a WasmEdge_Async execution.

WasmEdge_AsyncDeleteโ

Deletion of the WasmEdge_Async.

WasmEdge_AsyncGetโ

Wait and get the result of WasmEdge_Async execution.

WasmEdge_AsyncGetReturnsLengthโ

Wait and get the return list length of the WasmEdge_Async execution.

WasmEdge_AsyncWaitโ

Wait a WasmEdge_Async execution.

WasmEdge_AsyncWaitForโ

Wait a WasmEdge_Async execution with timeout.

WasmEdge_BytesCreateโ

Creation of the WasmEdge_Bytes with the buffer and its length.

WasmEdge_BytesDeleteโ

Deletion of the WasmEdge_Bytes.

WasmEdge_BytesWrapโ

Create the WasmEdge_Bytes wraps to the buffer.

WasmEdge_CallingFrameGetExecutorโ

Get the executor context from the current calling frame.

WasmEdge_CallingFrameGetMemoryInstanceโ

Get the memory instance by index from the module instance of the current calling frame.

WasmEdge_CallingFrameGetModuleInstanceโ

Get the module instance of the current calling frame.

WasmEdge_CompilerCompileโ

Compile the input WASM from the file path.

WasmEdge_CompilerCompileFromBufferโ

Compile the input WASM from the given buffer.

WasmEdge_CompilerCompileFromBytesโ

Compile the input WASM from a WasmEdge_Bytes.

WasmEdge_CompilerCreateโ

Creation of the WasmEdge_CompilerContext.

WasmEdge_CompilerDeleteโ

Deletion of the WasmEdge_CompilerContext.

WasmEdge_ConfigureAddHostRegistrationโ

Add a built-in host registration setting into WasmEdge_ConfigureContext.

WasmEdge_ConfigureAddProposalโ

Add a proposal setting into the WasmEdge_ConfigureContext.

WasmEdge_ConfigureCompilerGetOptimizationLevelโ

Get the optimization level of the AOT compiler.

WasmEdge_ConfigureCompilerGetOutputFormatโ

Get the output binary format of the AOT compiler.

WasmEdge_ConfigureCompilerIsDumpIRโ

Get the dump IR option of the AOT compiler.

WasmEdge_ConfigureCompilerIsGenericBinaryโ

Get the generic binary option of the AOT compiler.

WasmEdge_ConfigureCompilerIsInterruptibleโ

Get the interruptible option of the AOT compiler.

WasmEdge_ConfigureCompilerSetDumpIRโ

Set the dump IR option of the AOT compiler.

WasmEdge_ConfigureCompilerSetGenericBinaryโ

Set the generic binary option of the AOT compiler.

WasmEdge_ConfigureCompilerSetInterruptibleโ

Set the interruptible option of the AOT compiler.

WasmEdge_ConfigureCompilerSetOptimizationLevelโ

Set the optimization level of the AOT compiler.

WasmEdge_ConfigureCompilerSetOutputFormatโ

Set the output binary format of the AOT compiler.

WasmEdge_ConfigureCreateโ

Creation of the WasmEdge_ConfigureContext.

WasmEdge_ConfigureDeleteโ

Deletion of the WasmEdge_ConfigureContext.

WasmEdge_ConfigureGetMaxMemoryPageโ

Get the setting of the page limit of memory instances.

WasmEdge_ConfigureHasHostRegistrationโ

Check if a built-in host registration setting exists in the WasmEdge_ConfigureContext or not.

WasmEdge_ConfigureHasProposalโ

Check if a proposal setting exists in the WasmEdge_ConfigureContext or not.

WasmEdge_ConfigureIsAllowAFUNIXโ

Get the AllowAFUNIX option.

WasmEdge_ConfigureIsForceInterpreterโ

Get the force interpreter mode execution option.

WasmEdge_ConfigureRemoveHostRegistrationโ

Remove a built-in host registration setting in the WasmEdge_ConfigureContext.

WasmEdge_ConfigureRemoveProposalโ

Remove a proposal setting in the WasmEdge_ConfigureContext.

WasmEdge_ConfigureSetAllowAFUNIXโ

Set the option of enabling/disabling AF_UNIX support in the WASI socket.

WasmEdge_ConfigureSetForceInterpreterโ

Set the force interpreter mode execution option.

WasmEdge_ConfigureSetMaxMemoryPageโ

Set the page limit of memory instances.

WasmEdge_ConfigureStatisticsIsCostMeasuringโ

Get the cost measuring option for the statistics.

WasmEdge_ConfigureStatisticsIsInstructionCountingโ

Get the instruction counting option for the statistics.

WasmEdge_ConfigureStatisticsIsTimeMeasuringโ

Get the time measuring option for the statistics.

WasmEdge_ConfigureStatisticsSetCostMeasuringโ

Set the cost measuring option for the statistics.

WasmEdge_ConfigureStatisticsSetInstructionCountingโ

Set the instruction counting option for the statistics.

WasmEdge_ConfigureStatisticsSetTimeMeasuringโ

Set the time measuring option for the statistics.

WasmEdge_Driver_Compilerโ

Entrypoint for the compiler tool.

WasmEdge_Driver_Toolโ

Entrypoint for the runtime tool.

WasmEdge_Driver_UniToolโ

Entrypoint for the unified tool.

WasmEdge_ExecutorAsyncInvokeโ

Asynchronous invoke a WASM function by the function instance.

WasmEdge_ExecutorCreateโ

Creation of the WasmEdge_ExecutorContext.

WasmEdge_ExecutorDeleteโ

Deletion of the WasmEdge_ExecutorContext.

WasmEdge_ExecutorExperimentalRegisterPostHostFunctionโ

Register a host function that will be invoked after executing any host functions.

WasmEdge_ExecutorExperimentalRegisterPreHostFunctionโ

Register a host function that will be invoked before executing any host functions.

WasmEdge_ExecutorInstantiateโ

Instantiate an AST Module into a module instance.

WasmEdge_ExecutorInvokeโ

Invoke a WASM function by the function instance.

WasmEdge_ExecutorRegisterโ

Instantiate an AST Module into a named module instance and link into store.

WasmEdge_ExecutorRegisterImportโ

Register a module instance into a store with exporting its module name.

WasmEdge_ExportTypeGetExternalNameโ

Get the external name from an export type.

WasmEdge_ExportTypeGetExternalTypeโ

Get the external type from an export type.

WasmEdge_ExportTypeGetFunctionTypeโ

Get the external value (which is function type) from an export type.

WasmEdge_ExportTypeGetGlobalTypeโ

Get the external value (which is global type) from an export type.

WasmEdge_ExportTypeGetMemoryTypeโ

Get the external value (which is memory type) from an export type.

WasmEdge_ExportTypeGetTableTypeโ

Get the external value (which is table type) from an export type.

WasmEdge_ExportTypeGetTagTypeโ

Get the external value (which is tag type) from an export type.

WasmEdge_FunctionInstanceCreateโ

Creation of the WasmEdge_FunctionInstanceContext for host functions.

WasmEdge_FunctionInstanceCreateBindingโ

Creation of the WasmEdge_FunctionInstanceContext for host functions.

WasmEdge_FunctionInstanceDeleteโ

Deletion of the WasmEdge_FunctionInstanceContext.

WasmEdge_FunctionInstanceGetDataโ

Get the function data field of the function instance.

WasmEdge_FunctionInstanceGetFunctionTypeโ

Get the function type context of the function instance.

WasmEdge_FunctionTypeCreateโ

Creation of the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeDeleteโ

Deletion of the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetParametersโ

Get the parameter types list from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetParametersLengthโ

Get the parameter types list length from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetReturnsโ

Get the return types list from the WasmEdge_FunctionTypeContext.

WasmEdge_FunctionTypeGetReturnsLengthโ

Get the return types list length from the WasmEdge_FunctionTypeContext.

WasmEdge_GlobalInstanceCreateโ

Creation of the WasmEdge_GlobalInstanceContext.

WasmEdge_GlobalInstanceDeleteโ

Deletion of the WasmEdge_GlobalInstanceContext.

WasmEdge_GlobalInstanceGetGlobalTypeโ

Get the global type context from a global instance.

WasmEdge_GlobalInstanceGetValueโ

Get the value from a global instance.

WasmEdge_GlobalInstanceSetValueโ

Set the value into a global instance.

WasmEdge_GlobalTypeCreateโ

Creation of the WasmEdge_GlobalTypeContext.

WasmEdge_GlobalTypeDeleteโ

Deletion of the WasmEdge_GlobalTypeContext.

WasmEdge_GlobalTypeGetMutabilityโ

Get the mutability from a global type.

WasmEdge_GlobalTypeGetValTypeโ

Get the value type from a global type.

WasmEdge_ImportTypeGetExternalNameโ

Get the external name from an import type.

WasmEdge_ImportTypeGetExternalTypeโ

Get the external type from an import type.

WasmEdge_ImportTypeGetFunctionTypeโ

Get the external value (which is function type) from an import type.

WasmEdge_ImportTypeGetGlobalTypeโ

Get the external value (which is global type) from an import type.

WasmEdge_ImportTypeGetMemoryTypeโ

Get the external value (which is memory type) from an import type.

WasmEdge_ImportTypeGetModuleNameโ

Get the module name from an import type.

WasmEdge_ImportTypeGetTableTypeโ

Get the external value (which is table type) from an import type.

WasmEdge_ImportTypeGetTagTypeโ

Get the external value (which is tag type) from an import type.

WasmEdge_LimitIsEqualโ

Compare the two WasmEdge_Limit objects.

WasmEdge_LoaderCreateโ

Creation of the WasmEdge_LoaderContext.

WasmEdge_LoaderDeleteโ

Deletion of the WasmEdge_LoaderContext.

WasmEdge_LoaderParseFromBufferโ

Load and parse the WASM module from a buffer into WasmEdge_ASTModuleContext.

WasmEdge_LoaderParseFromBytesโ

Load and parse the WASM module from a WasmEdge_Bytes into WasmEdge_ASTModuleContext.

WasmEdge_LoaderParseFromFileโ

Load and parse the WASM module from a WASM file into a WasmEdge_ASTModuleContext.

WasmEdge_LoaderSerializeASTModuleโ

Serialize the WasmEdge_ASTModuleContext into WASM binary.

WasmEdge_LogOffโ

Set the logging system off.

WasmEdge_LogSetDebugLevelโ

Set the logging system to filter to debug level.

WasmEdge_LogSetErrorLevelโ

Set the logging system to filter to error level.

WasmEdge_MemoryInstanceCreateโ

Creation of the WasmEdge_MemoryInstanceContext.

WasmEdge_MemoryInstanceDeleteโ

Deletion of the WasmEdge_MemoryInstanceContext.

WasmEdge_MemoryInstanceGetDataโ

Copy the data to the output buffer from a memory instance.

WasmEdge_MemoryInstanceGetMemoryTypeโ

Get the memory type context from a memory instance.

WasmEdge_MemoryInstanceGetPageSizeโ

Get the current page size (64 KiB of each page) of a memory instance.

WasmEdge_MemoryInstanceGetPointerโ

Get the data pointer in a memory instance.

WasmEdge_MemoryInstanceGetPointerConstโ

Get the const data pointer in a const memory instance.

WasmEdge_MemoryInstanceGrowPageโ

Grow a memory instance with a page size.

WasmEdge_MemoryInstanceSetDataโ

Copy the data into a memory instance from the input buffer.

WasmEdge_MemoryTypeCreateโ

Creation of the WasmEdge_MemoryTypeContext.

WasmEdge_MemoryTypeDeleteโ

Deletion of the WasmEdge_MemoryTypeContext.

WasmEdge_MemoryTypeGetLimitโ

Get the limit from a memory type.

WasmEdge_ModuleInstanceAddFunctionโ

Add a function instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddGlobalโ

Add a global instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddMemoryโ

Add a memory instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceAddTableโ

Add a table instance context into a WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceCreateโ

Creation of the WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceCreateWASIโ

Creation of the WasmEdge_ModuleInstanceContext for the WASI specification.

WasmEdge_ModuleInstanceCreateWithDataโ

Creation of the WasmEdge_ModuleInstanceContext with host data.

WasmEdge_ModuleInstanceDeleteโ

Deletion of the WasmEdge_ModuleInstanceContext.

WasmEdge_ModuleInstanceFindFunctionโ

Get the exported function instance context of a module instance.

WasmEdge_ModuleInstanceFindGlobalโ

Get the exported global instance context of a module instance.

WasmEdge_ModuleInstanceFindMemoryโ

Get the exported memory instance context of a module instance.

WasmEdge_ModuleInstanceFindTableโ

Get the exported table instance context of a module instance.

WasmEdge_ModuleInstanceFindTagโ

Get the exported tag instance context of a module instance.

WasmEdge_ModuleInstanceGetHostDataโ

Get the host data set into the module instance when creating.

WasmEdge_ModuleInstanceGetModuleNameโ

Get the export module name of a module instance.

WasmEdge_ModuleInstanceInitWASIโ

Initialize the WasmEdge_ModuleInstanceContext for the WASI specification.

WasmEdge_ModuleInstanceInitWasmEdgeProcessโ

Initialize the WasmEdge_ModuleInstanceContext for the wasmedge_process specification.

WasmEdge_ModuleInstanceListFunctionโ

List the exported function names of a module instance.

WasmEdge_ModuleInstanceListFunctionLengthโ

Get the length of exported function list of a module instance.

WasmEdge_ModuleInstanceListGlobalโ

List the exported global names of a module instance.

WasmEdge_ModuleInstanceListGlobalLengthโ

Get the length of exported global list of a module instance.

WasmEdge_ModuleInstanceListMemoryโ

List the exported memory names of a module instance.

WasmEdge_ModuleInstanceListMemoryLengthโ

Get the length of exported memory list of a module instance.

WasmEdge_ModuleInstanceListTableโ

List the exported table names of a module instance.

WasmEdge_ModuleInstanceListTableLengthโ

Get the length of exported table list of a module instance.

WasmEdge_ModuleInstanceListTagโ

List the exported tag names of a module instance.

WasmEdge_ModuleInstanceListTagLengthโ

Get the length of exported tag list of a module instance.

WasmEdge_ModuleInstanceWASIGetExitCodeโ

Get the WASI exit code.

WasmEdge_ModuleInstanceWASIGetNativeHandlerโ

Get the native handler from the WASI mapped FD/Handler.

WasmEdge_PluginCreateModuleโ

Create the module instance in the plug-in by the module name.

WasmEdge_PluginFindโ

Find the loaded plug-in context by name.

WasmEdge_PluginGetPluginNameโ

Get the plug-in name of the plug-in context.

WasmEdge_PluginInitWASINNโ

Initialize the wasi_nn plug-in.

WasmEdge_PluginListModuleโ

List the modules in the plug-in context with their names.

WasmEdge_PluginListModuleLengthโ

Get the length of module list in the plug-in context.

WasmEdge_PluginListPluginsโ

List the loaded plug-ins with their names.

WasmEdge_PluginListPluginsLengthโ

Get the length of loaded plug-in list.

WasmEdge_PluginLoadFromPathโ

Load the plugin with the given file or directory.

WasmEdge_PluginLoadWithDefaultPathsโ

Load plugins with the default search paths.

WasmEdge_Plugin_GetDescriptorโ

Implement by plugins for returning the plugin descriptor.

WasmEdge_ResultGenโ

Generate the result with code.

WasmEdge_ResultGetCategoryโ

Get the error category.

WasmEdge_ResultGetCodeโ

Get the result code.

WasmEdge_ResultGetMessageโ

Get the result message.

WasmEdge_ResultOKโ

Check the result is a success or not.

WasmEdge_StatisticsClearโ

Clear all data in the WasmEdge_StatisticsContext.

WasmEdge_StatisticsCreateโ

Creation of the WasmEdge_StatisticsContext.

WasmEdge_StatisticsDeleteโ

Deletion of the WasmEdge_StatisticsContext.

WasmEdge_StatisticsGetInstrCountโ

Get the instruction count in execution.

WasmEdge_StatisticsGetInstrPerSecondโ

Get the instruction count per second in execution.

WasmEdge_StatisticsGetTotalCostโ

Get the total cost in execution.

WasmEdge_StatisticsSetCostLimitโ

Set the cost limit in execution.

WasmEdge_StatisticsSetCostTableโ

Set the costs of instructions.

WasmEdge_StoreCreateโ

Creation of the WasmEdge_StoreContext.

WasmEdge_StoreDeleteโ

Deletion of the WasmEdge_StoreContext.

WasmEdge_StoreFindModuleโ

Get the module instance context by the module name.

WasmEdge_StoreListModuleโ

List the registered module names.

WasmEdge_StoreListModuleLengthโ

Get the length of registered module list in store.

WasmEdge_StringCopyโ

Copy the content of WasmEdge_String object to the buffer.

WasmEdge_StringCreateByBufferโ

Creation of the WasmEdge_String with the buffer and its length.

WasmEdge_StringCreateByCStringโ

Creation of the WasmEdge_String with the C string.

WasmEdge_StringDeleteโ

Deletion of the WasmEdge_String.

WasmEdge_StringIsEqualโ

Compare the two WasmEdge_String objects.

WasmEdge_StringWrapโ

Create the WasmEdge_String wraps to the buffer.

WasmEdge_TableInstanceCreateโ

Creation of the WasmEdge_TableInstanceContext.

WasmEdge_TableInstanceCreateWithInitโ

Creation of the WasmEdge_TableInstanceContext with the default init value.

WasmEdge_TableInstanceDeleteโ

Deletion of the WasmEdge_TableInstanceContext.

WasmEdge_TableInstanceGetDataโ

Get the reference value in a table instance.

WasmEdge_TableInstanceGetSizeโ

Get the size of a table instance.

WasmEdge_TableInstanceGetTableTypeโ

Get the table type context from a table instance.

WasmEdge_TableInstanceGrowโ

Grow a table instance with a size.

WasmEdge_TableInstanceSetDataโ

Set the reference value into a table instance.

WasmEdge_TableTypeCreateโ

Creation of the WasmEdge_TableTypeContext.

WasmEdge_TableTypeDeleteโ

Deletion of the WasmEdge_TableTypeContext.

WasmEdge_TableTypeGetLimitโ

Get the limit from a table type.

WasmEdge_TableTypeGetRefTypeโ

Get the reference type from a table type.

WasmEdge_TagInstanceGetTagTypeโ

Get the tag type context from a tag instance.

WasmEdge_TagTypeGetFunctionTypeโ

Get the function type from a tag type.

WasmEdge_VMAsyncExecuteโ

Asynchronous invoke a WASM function by name.

WasmEdge_VMAsyncExecuteRegisteredโ

Asynchronous invoke a WASM function by its module name and function name.

WasmEdge_VMAsyncRunWasmFromASTModuleโ

Instantiate the WASM module from a WasmEdge AST Module and asynchronous invoke a function by name.

WasmEdge_VMAsyncRunWasmFromBufferโ

Instantiate the WASM module from a buffer and asynchronous invoke a function by name.

WasmEdge_VMAsyncRunWasmFromBytesโ

Instantiate the WASM module from a WasmEdge_Bytes and asynchronous invoke a function by name.

WasmEdge_VMAsyncRunWasmFromFileโ

Instantiate the WASM module from a WASM file and asynchronous invoke a function by name.

WasmEdge_VMCleanupโ

Reset of WasmEdge_VMContext.

WasmEdge_VMCreateโ

Creation of the WasmEdge_VMContext.

WasmEdge_VMDeleteโ

Deletion of the WasmEdge_VMContext.

WasmEdge_VMExecuteโ

Invoke a WASM function by name.

WasmEdge_VMExecuteRegisteredโ

Invoke a WASM function by its module name and function name.

WasmEdge_VMGetActiveModuleโ

Get the current instantiated module in VM.

WasmEdge_VMGetExecutorContextโ

Get the executor context used in the WasmEdge_VMContext.

WasmEdge_VMGetFunctionListโ

Get the exported function list.

WasmEdge_VMGetFunctionListLengthโ

Get the length of exported function list.

WasmEdge_VMGetFunctionTypeโ

Get the function type by function name.

WasmEdge_VMGetFunctionTypeRegisteredโ

Get the function type by function name.

WasmEdge_VMGetImportModuleContextโ

Get the module instance corresponding to the WasmEdge_HostRegistration settings.

WasmEdge_VMGetLoaderContextโ

Get the loader context used in the WasmEdge_VMContext.

WasmEdge_VMGetRegisteredModuleโ

Get the registered module in VM by the module name.

WasmEdge_VMGetStatisticsContextโ

Get the statistics context used in the WasmEdge_VMContext.

WasmEdge_VMGetStoreContextโ

Get the store context used in the WasmEdge_VMContext.

WasmEdge_VMGetValidatorContextโ

Get the validator context used in the WasmEdge_VMContext.

WasmEdge_VMInstantiateโ

Instantiate the validated WASM module in the VM context.

WasmEdge_VMListRegisteredModuleโ

List the registered module names in the WasmEdge_VMContext.

WasmEdge_VMListRegisteredModuleLengthโ

Get the length of registered module list in the WasmEdge_VMContext.

WasmEdge_VMLoadWasmFromASTModuleโ

Load the WASM module from loaded WasmEdge AST Module.

WasmEdge_VMLoadWasmFromBufferโ

Load the WASM module from a buffer.

WasmEdge_VMLoadWasmFromBytesโ

Load the WASM module from a WasmEdge_Bytes.

WasmEdge_VMLoadWasmFromFileโ

Load the WASM module from a WASM file.

WasmEdge_VMRegisterModuleFromASTModuleโ

Instantiate and register an AST Module into a named module instance in VM.

WasmEdge_VMRegisterModuleFromBufferโ

Register and instantiate WASM into the store in VM from a buffer.

WasmEdge_VMRegisterModuleFromBytesโ

Register and instantiate WASM into the store in VM from a WasmEdge_Bytes.

WasmEdge_VMRegisterModuleFromFileโ

Register and instantiate WASM into the store in VM from a WASM file.

WasmEdge_VMRegisterModuleFromImportโ

Register a module instance into the store in VM with exporting its module name.

WasmEdge_VMRunWasmFromASTModuleโ

Instantiate the WASM module from a WasmEdge AST Module and invoke a function by name.

WasmEdge_VMRunWasmFromBufferโ

Instantiate the WASM module from a buffer and invoke a function by name.

WasmEdge_VMRunWasmFromBytesโ

Instantiate the WASM module from a WasmEdge_Bytes and invoke a function by name.

WasmEdge_VMRunWasmFromFileโ

Instantiate the WASM module from a WASM file and invoke a function by name.

WasmEdge_VMValidateโ

Validate the WASM module loaded into the VM context.

WasmEdge_ValTypeGenExternRefโ

Generate the ExternRef WASM value type.

WasmEdge_ValTypeGenF32โ

Generate the F32 WASM value type.

WasmEdge_ValTypeGenF64โ

Generate the F64 WASM value type.

WasmEdge_ValTypeGenFuncRefโ

Generate the FuncRef WASM value type.

WasmEdge_ValTypeGenI32โ

Generate the I32 WASM value type.

WasmEdge_ValTypeGenI64โ

Generate the I64 WASM value type.

WasmEdge_ValTypeGenV128โ

Generate the V128 WASM value type.

WasmEdge_ValTypeIsEqualโ

Compare the two WasmEdge_ValType objects.

WasmEdge_ValTypeIsExternRefโ

Specify the WASM value type is an ExternRef or not.

WasmEdge_ValTypeIsF32โ

Specify the WASM value type is a F32 or not.

WasmEdge_ValTypeIsF64โ

Specify the WASM value type is a F64 or not.

WasmEdge_ValTypeIsFuncRefโ

Specify the WASM value type is a FuncRef or not.

WasmEdge_ValTypeIsI32โ

Specify the WASM value type is an I32 or not.

WasmEdge_ValTypeIsI64โ

Specify the WASM value type is an I64 or not.

WasmEdge_ValTypeIsRefโ

Specify the WASM value type is a Ref (includes nullable and non-nullable) or not.

WasmEdge_ValTypeIsRefNullโ

Specify the WASM value type is a nullable Ref or not.

WasmEdge_ValTypeIsV128โ

Specify the WASM value type is a V128 or not.

WasmEdge_ValidatorCreateโ

Creation of the WasmEdge_ValidatorContext.

WasmEdge_ValidatorDeleteโ

Deletion of the WasmEdge_ValidatorContext.

WasmEdge_ValidatorValidateโ

Validate the WasmEdge AST Module.

WasmEdge_ValueGenExternRefโ

Generate the function reference WASM value.

WasmEdge_ValueGenF32โ

Generate the F32 WASM value.

WasmEdge_ValueGenF64โ

Generate the F64 WASM value.

WasmEdge_ValueGenFuncRefโ

Generate the function reference WASM value.

WasmEdge_ValueGenI32โ

Generate the I32 WASM value.

WasmEdge_ValueGenI64โ

Generate the I64 WASM value.

WasmEdge_ValueGenV128โ

Generate the V128 WASM value.

WasmEdge_ValueGetExternRefโ

Retrieve the external reference from the WASM value.

WasmEdge_ValueGetF32โ

Retrieve the F32 value from the WASM value.

WasmEdge_ValueGetF64โ

Retrieve the F64 value from the WASM value.

WasmEdge_ValueGetFuncRefโ

Retrieve the function instance context from the WASM value.

WasmEdge_ValueGetI32โ

Retrieve the I32 value from the WASM value.

WasmEdge_ValueGetI64โ

Retrieve the I64 value from the WASM value.

WasmEdge_ValueGetV128โ

Retrieve the V128 value from the WASM value.

WasmEdge_ValueIsNullRefโ

Specify the WASM value is a null reference or not.

WasmEdge_VersionGetโ

Get the version string of the WasmEdge C API.

WasmEdge_VersionGetMajorโ

Get the major version value of the WasmEdge C API.

WasmEdge_VersionGetMinorโ

Get the minor version value of the WasmEdge C API.

WasmEdge_VersionGetPatchโ

Get the patch version value of the WasmEdge C API.

Type Aliases

WasmEdge_CompilerOptimizationLevel

WasmEdge_CompilerOutputFormat

WasmEdge_ErrCategory

WasmEdge_ErrCode

WasmEdge_ExternalType

WasmEdge_HostFunc_t

WasmEdge_HostRegistration

WasmEdge_Mutability

WasmEdge_ProgramOptionType

Type of option value.

WasmEdge_Proposal

WasmEdge_TypeCode

WasmEdge_WrapFunc_t

int128_t

uint128_t