wasmedge_sys::instance::memory
Struct Memory

pub struct Memory { /* private fields */ }

Defines a WebAssembly memory instance, which is a linear memory described by its type. Each memory instance consists of a vector of bytes and an optional maximum size, and its size is a multiple of the WebAssembly page size (64KiB of each page).

Implementations

impl Memory

pub fn create(ty: &MemoryType) -> WasmEdgeResult<Self>

Create a new Memory to be associated with the given capacity limit.

Arguments

ty - The type of the new Memory instance.

Errors

If fail to create the memory instance, then WasmEdgeError::Mem(MemError::Create) is returned.

pub fn ty(&self) -> WasmEdgeResult<MemoryType>

Returns the type of the Memory.

Errors

If fail to get the type from the Memory, then an error is returned.

pub fn get_data(&self, offset: u32, len: u32) -> WasmEdgeResult<Vec<u8>>

Copies the data from the Memory to the output buffer.

Arguments

offset - The data start offset in the Memory.

len - The requested data length.

Errors

If the offset + len is larger than the data size in the Memory, then an error is returned.

pub fn set_data( &mut self, data: impl AsRef<[u8]>, offset: u32, ) -> WasmEdgeResult<()>

Copies the data from the given input buffer into the Memory.

Arguments

data - The data buffer to copy.

offset - The data start offset in the Memory.

Errors

If the sum of the offset and the data length is larger than the size of the Memory, then an error is returned.

pub unsafe fn data_pointer( &self, offset: u32, len: u32, ) -> WasmEdgeResult<*const u8>

Returns the const data pointer to the Memory.

Arguments

offset - The data start offset in the Memory.

len - The requested data length. If the size of offset + len is larger than the data size in the Memory

Errors

If fail to get the data pointer, then an error is returned.

Safety

The lifetime of the returned pointer must not exceed that of the object itself.

pub unsafe fn data_pointer_mut( &mut self, offset: u32, len: u32, ) -> WasmEdgeResult<*mut u8>

Returns the data pointer to the Memory.

Arguments

offset - The data start offset in the Memory.

len - The requested data length. If the size of offset + len is larger than the data size in the Memory

Errors

If fail to get the data pointer, then an error is returned.

Safety

The lifetime of the returned pointer must not exceed that of the object itself.

pub fn size(&self) -> u32

Returns the size, in WebAssembly pages (64 KiB of each page), of this wasm memory.

pub fn grow(&mut self, count: u32) -> WasmEdgeResult<()>

Grows this WebAssembly memory by count pages.

Arguments

count - The page counts to be extended to the Memory.

Errors

If fail to grow the page count, then an error is returned.

pub unsafe fn as_ptr(&self) -> *const WasmEdge_MemoryInstanceContext

Safety

Provides a raw pointer to the inner memory context. The lifetime of the returned pointer must not exceed that of the object itself.

impl Memory

pub fn get_ref<T: Sized>(&self, offset: usize) -> Option<&T>

pub fn slice<T: Sized>(&self, offset: usize, len: usize) -> Option<&[T]>

pub fn get_ref_mut<T: Sized>(&mut self, offset: usize) -> Option<&mut T>

pub fn mut_slice<T: Sized>(&self, offset: usize, len: usize) -> Option<&mut [T]>

pub fn write<T: Sized>(&mut self, offset: usize, data: T) -> Option<()>