Struct wasmedge_sys::Config

pub struct Config { /* private fields */ }

Defines Config struct used to check/set the configuration options.

Config manages the configuration options, which are used to initiate WasmEdge Vm, Loader, Validator, Executor, and Compiler.

The configuration options are categorized into the following four groups:

WebAssembly Proposals

This group of options are used to turn on/off the WebAssembly proposals. They are effective to any WasmEdge context created with Config.

MultiMemories enables to use multiple memories within a single Wasm module.

Also see Multiple Memories for Wasm

ImportExportMutGlobals supports mutable imported and exported globals.

Also see Import/Export Mutable Globals Proposal.

NonTrapFloatToIntConversions supports the non-trapping float-to-int conversion.

Also see Non-trapping Float-to-int Conversions Proposal.

SignExtensionOperators supports new integer instructions for sign-extending 8-bit, 16-bit, and 32-bit values.

Also see Sign-extension Operators Proposal.

MultiValue supports functions and instructions with multiple return values, and blocks with inputs.

Also see Multi-value Extension.

BulkMemoryOperations supports bulk memory operations.

Also see Bulk Memory Operations Proposal.

ReferenceTypes supports reference types.

Also see Reference Types Proposal.

SIMD supports 128-bit packed SIMD extension to WebAssembly.

Also see SIMD Proposal.

TailCall supports tail call optimization.

Also see Tail Call Proposal.

Annotations supports annotations in WASM text format.

Also see Annotations Proposal.

Memory64 supports 64-bit memory indexes.

Also see Memory64 Proposal.

Threads supports threading feature.

Also see Threading Proposal.

ExceptionHandling supports exception handling.

Also see Exception Handling Proposal.

FunctionReferences supports typed function references for WebAssembly.

Also see Function References Proposal.

Host Registrations

Wasi turns on the WASI support in Vm.

WasmEdgeProcess turns on the wasmedge_process support in Vm.

The two options are only effective to Vm.

Memory Management

maximum_memory_page limits the page size of Memory. This option is only effective to Executor and Vm.

AOT Compilation

The AOT compiler options configure the behavior about optimization level, output format, dump IR, and generic binary.

Compiler Optimization Levels

O0 performs as many optimizations as possible.

O1 optimizes quickly without destroying debuggability.

02 optimizes for fast execution as much as possible without triggering significant incremental compile time or code size growth.

O3 optimizes for fast execution as much as possible.

Os optimizes for small code size as much as possible without triggering significant incremental compile time or execution time slowdowns.

Oz optimizes for small code size as much as possible.

Compiler Output Formats

Native specifies the output format is native dynamic library (*.wasm.so).

Wasm specifies the output format is WebAssembly with AOT compiled codes in custom section (*.wasm).

dump_ir determines if AOT compiler generates IR or not.

generic_binary determines if AOT compiler generates the generic binary or not.

interruptible determines if AOT compiler generates interruptible binary or not.

The configuration options above are only effective to Compiler.

Runtime Statistics

instr_counting determines if measuring the count of instructions when running a compiled or pure WASM.

cost_measuring determines if measuring the instruction costs when running a compiled or pure WASM.

time_measuring determines if measuring the running time when running a compiled or pure WASM.

API users can first set the options of interest, such as those related to the WebAssembly proposals, host registrations, AOT compiler options, and etc., then apply the configuration to create other WasmEdge runtime structs.

Implementations

impl Config

pub fn create() -> WasmEdgeResult<Self>

Creates a new Config.

Error

If fail to create, then an error is returned.

pub fn wasi(&mut self, enable: bool)

Enables or disables host registration wasi. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn wasi_enabled(&self) -> bool

Checks if host registration wasi turns on or not.

pub fn wasmedge_process(&mut self, enable: bool)

Enables or disables host registration WasmEdge process. By default, the option is disabled.

Notice that to enable the wasmege_process option in Vm, it MUST be guaranteed that the wasmedge_process plugins are loaded first. If not, use the load_plugin_from_default_paths function to load the relevant plugins from the default paths

Argument

enable - Whether the option turns on or not.

pub fn wasmedge_process_enabled(&self) -> bool

Checks if host registration wasmedge process turns on or not.

pub fn set_max_memory_pages(&mut self, count: u32)

Sets the maximum number of the memory pages available.

Argument

count - The page count (64KB per page).

pub fn get_max_memory_pages(&self) -> u32

Returns the number of the memory pages available.

pub fn mutable_globals(&mut self, enable: bool)

Enables or disables the ImportExportMutGlobals option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn mutable_globals_enabled(&self) -> bool

Checks if the ImportExportMutGlobals option turns on or not.

pub fn non_trap_conversions(&mut self, enable: bool)

Enables or disables the NonTrapFloatToIntConversions option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn non_trap_conversions_enabled(&self) -> bool

Checks if the NonTrapFloatToIntConversions option turns on or not.

pub fn sign_extension_operators(&mut self, enable: bool)

Enables or disables the SignExtensionOperators option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn sign_extension_operators_enabled(&self) -> bool

Checks if the SignExtensionOperators option turns on or not.

pub fn multi_value(&mut self, enable: bool)

Enables or disables the MultiValue option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn multi_value_enabled(&self) -> bool

Checks if the MultiValue option turns on or not.

pub fn bulk_memory_operations(&mut self, enable: bool)

Enables or disables the BulkMemoryOperations option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn bulk_memory_operations_enabled(&self) -> bool

Checks if the BulkMemoryOperations option turns on or not.

pub fn reference_types(&mut self, enable: bool)

Enables or disables the ReferenceTypes option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn reference_types_enabled(&self) -> bool

Checks if the ReferenceTypes option turns on or not.

pub fn simd(&mut self, enable: bool)

Enables or disables the SIMD option. By default, the option is enabled.

Argument

enable - Whether the option turns on or not.

pub fn simd_enabled(&self) -> bool

Checks if the SIMD option turns on or not.

pub fn tail_call(&mut self, enable: bool)

Enables or disables the TailCall option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn tail_call_enabled(&self) -> bool

Checks if the TailCall option turns on or not.

pub fn annotations(&mut self, enable: bool)

Enables or disables the Annotations option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn annotations_enabled(&self) -> bool

Checks if the Annotations option turns on or not.

pub fn memory64(&mut self, enable: bool)

Enables or disables the Memory64 option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn memory64_enabled(&self) -> bool

Checks if the Memory64 option turns on or not.

pub fn threads(&mut self, enable: bool)

Enables or disables the Threads option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn threads_enabled(&self) -> bool

Checks if the Threads option turns on or not.

pub fn exception_handling(&mut self, enable: bool)

Enables or disables the ExceptionHandling option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn exception_handling_enabled(&self) -> bool

Checks if the ExceptionHandling option turns on or not.

pub fn function_references(&mut self, enable: bool)

Enables or disables the FunctionReferences option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn function_references_enabled(&self) -> bool

Checks if the FunctionReferences option turns on or not.

pub fn multi_memories(&mut self, enable: bool)

Enables or disables the MultiMemories option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn multi_memories_enabled(&self) -> bool

Checks if the MultiMemories option turns on or not.

pub fn interpreter_mode(&mut self, enable: bool)

Enables or disables the ForceInterpreter option. By default, the option is disabled.

Argument

enable - Whether the option turns on or not.

pub fn interpreter_mode_enabled(&self) -> bool

Checks if the ForceInterpreter option turns on or not.

pub fn count_instructions(&mut self, flag: bool)

Sets the instruction counting option. By default, the option is disabled.

Argument

flag - Whether support instruction counting or not when execution after AOT compilation.

pub fn is_instruction_counting(&self) -> bool

Checks if the instruction counting option turns on or not.

pub fn measure_cost(&mut self, flag: bool)

Sets the cost measuring option. By default, the option is disabled.

Argument

flag - Whether support cost measuring or not when execution after AOT compilation.

pub fn is_cost_measuring(&self) -> bool

Checks if the cost measuring option turns on or not.

pub fn measure_time(&mut self, flag: bool)

Sets the time measuring option. By default, the option is disabled.

Argument

flag - Whether support time measuring or not when execution after AOT compilation.

pub fn is_time_measuring(&self) -> bool

Checks if the time measuring option turns on or not.